@@ -1,72 +1,81 @@
 {
-  "name": "date-time-picker-app",
-  "version": "0.0.0",
+  "name": "@danielmoncada/angular-datetime-picker",
+  "version": "17.0.0",
   "description": "Angular Date Time Picker",
   "keywords": [
     "Angular",
     "datepicker",
     "date picker",
     "timepicker",
-    "time picker"
+    "time picker",
+    "datetime picker",
+    "date time picker",
+    "material",
+    "ngx"
   ],
-  "author": "daniel pan <myron.yk.pan@gmail.com>",
+  "author": "Maintained and updated by Daniel Moncada, original implementatiom by Daniel Pan",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build picker",
     "test": "ng test picker",
+    "test-with-coverage": "ng test --watch=false --no-progress --browsers=ChromeHeadlessNoSandbox --code-coverage=true picker",
     "lint": "ng lint",
     "e2e": "ng e2e",
-    "build_lib": "ng build picker && cp README.md dist/picker",
-    "build_css": "mkdir -p dist/picker/assets/style && node-sass --output-style=compressed projects/picker/src/sass/picker.scss > dist/picker/assets/style/picker.min.css",
+    "build_lib_onWindows": "ng build picker --configuration production && copy \"README.md\" \"dist/picker\"",
+    "build_lib_onLinux": "ng build picker --configuration production && cp README.md dist/picker",
+    "build_css_onWindows": "mkdir \"dist/picker/assets/style\" && node-sass --output-style=compressed projects/picker/src/sass/picker.scss > dist/picker/assets/style/picker.min.css",
+    "build_css_onLinux": "mkdir -p dist/picker/assets/style && node-sass --output-style=compressed projects/picker/src/sass/picker.scss > dist/picker/assets/style/picker.min.css",
     "npm_pack": "cd dist/picker && npm pack",
-    "package": "npm run build_lib && npm run build_css && npm run npm_pack"
+    "package_windows": "npm run build_lib_onWindows && npm run build_css_onWindows && npm run npm_pack",
+    "package_linux": "npm run build_lib_onLinux && npm run build_css_onLinux && npm run npm_pack"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/DanielYKPan/date-time-picker.git"
+    "url": "https://github.com/danielmoncada/date-time-picker.git"
   },
   "dependencies": {
-    "@angular/animations": "~8.2.6",
-    "@angular/common": "~8.2.6",
-    "@angular/compiler": "~8.2.6",
-    "@angular/core": "~8.2.6",
-    "@angular/forms": "~8.2.6",
-    "@angular/platform-browser": "~8.2.6",
-    "@angular/platform-browser-dynamic": "~8.2.6",
-    "@angular/router": "~8.2.6",
-    "core-js": "^2.6.5",
-    "moment": "^2.24.0",
-    "moment-timezone": "^0.5.25",
-    "rxjs": "~6.5.3",
-    "zone.js": "~0.9.1"
+    "@angular/animations": "^17.0.4",
+    "@angular/common": "^17.0.4",
+    "@angular/compiler": "^17.0.4",
+    "@angular/core": "^17.0.4",
+    "@angular/forms": "^17.0.4",
+    "@angular/platform-browser": "^17.0.4",
+    "@angular/platform-browser-dynamic": "^17.0.4",
+    "@angular/router": "^17.0.4",
+    "core-js": "^3.22.3",
+    "moment": "^2.29.4",
+    "moment-timezone": "^0.5.43",
+    "np": "^8.0.4",
+    "rxjs": "^7.8.1",
+    "tslib": "^2.4.0",
+    "uuid": "^9.0.1",
+    "zone.js": "^0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.4",
-    "@angular-devkit/build-ng-packagr": "~0.803.4",
-    "@angular/cdk": "^7.3.7",
-    "@angular/cli": "~8.3.4",
-    "@angular/compiler-cli": "~8.2.6",
-    "@angular/language-service": "~8.2.6",
-    "@types/jasmine": "~3.3.12",
-    "@types/jasminewd2": "~2.0.6",
-    "@types/node": "~11.13.7",
-    "codelyzer": "^5.0.1",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.5",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "ng-packagr": "^5.4.0",
-    "node-sass": "^4.12.0",
-    "protractor": "~5.4.2",
-    "ts-node": "~8.1.0",
-    "tsickle": "^0.37.0",
-    "tslib": "^1.10.0",
-    "tslint": "~5.16.0",
-    "typescript": "~3.5.3"
+    "@angular-devkit/build-angular": "^17.0.3",
+    "@angular/cdk": "^17.0.1",
+    "@angular/cli": "^17.0.3",
+    "@angular/compiler-cli": "^17.0.4",
+    "@angular/language-service": "^17.0.4",
+    "@types/jasmine": "^3.10.6",
+    "@types/jasminewd2": "^2.0.10",
+    "@types/node": "^20.10.0",
+    "codecov": "^3.8.3",
+    "codelyzer": "^6.0.2",
+    "jasmine-core": "~4.1.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.3.10",
+    "karma-chrome-launcher": "^3.1.1",
+    "karma-coverage": "^2.2.0",
+    "karma-jasmine": "~5.0.0",
+    "karma-jasmine-html-reporter": "^1.7.0",
+    "ng-packagr": "^17.0.2",
+    "node-sass": "^9.0.0",
+    "protractor": "~7.0.0",
+    "ts-node": "^10.9.1",
+    "tslint": "^6.1.3",
+    "typescript": "~5.2.2"
   }
 }
