@@ -18,15 +18,12 @@ import {
     Optional,
     Output
 } from '@angular/core';
-import { OwlDateTimeIntl } from './date-time-picker-intl.service';
-import { DateTimeAdapter } from './adapter/date-time-adapter.class';
-import {
-    OWL_DATE_TIME_FORMATS,
-    OwlDateTimeFormats
-} from './adapter/date-time-format.class';
-import { SelectMode } from './date-time.class';
-import { take } from 'rxjs/operators';
-import { Subscription } from 'rxjs';
+import {OwlDateTimeIntl} from './date-time-picker-intl.service';
+import {DateTimeAdapter} from './adapter/date-time-adapter.class';
+import {OWL_DATE_TIME_FORMATS, OwlDateTimeFormats} from './adapter/date-time-format.class';
+import {DateView, DateViewType, SelectMode} from './date-time.class';
+import {take} from 'rxjs/operators';
+import {Subscription} from 'rxjs';
 
 @Component({
     selector: 'owl-date-time-calendar',
@@ -41,20 +38,9 @@ import { Subscription } from 'rxjs';
 })
 export class OwlCalendarComponent<T>
     implements OnInit, AfterContentInit, AfterViewChecked, OnDestroy {
-    /**
-     * Date filter for the month and year view
-     * */
-    @Input()
-    dateFilter: Function;
 
-    /**
-     * Set the first day of week
-     */
-    @Input()
-    firstDayOfWeek = 0;
+    DateView = DateView;
 
-    /** The minimum selectable date. */
-    private _minDate: T | null;
     @Input()
     get minDate(): T | null {
         return this._minDate;
@@ -66,15 +52,13 @@ export class OwlCalendarComponent<T>
 
         this._minDate = value
             ? this.dateTimeAdapter.createDate(
-                  this.dateTimeAdapter.getYear(value),
-                  this.dateTimeAdapter.getMonth(value),
-                  this.dateTimeAdapter.getDate(value)
-              )
+                this.dateTimeAdapter.getYear(value),
+                this.dateTimeAdapter.getMonth(value),
+                this.dateTimeAdapter.getDate(value)
+            )
             : null;
     }
 
-    /** The maximum selectable date. */
-    private _maxDate: T | null;
     @Input()
     get maxDate(): T | null {
         return this._maxDate;
@@ -86,15 +70,13 @@ export class OwlCalendarComponent<T>
 
         this._maxDate = value
             ? this.dateTimeAdapter.createDate(
-                  this.dateTimeAdapter.getYear(value),
-                  this.dateTimeAdapter.getMonth(value),
-                  this.dateTimeAdapter.getDate(value)
-              )
+                this.dateTimeAdapter.getYear(value),
+                this.dateTimeAdapter.getMonth(value),
+                this.dateTimeAdapter.getDate(value)
+            )
             : null;
     }
 
-    /** The current picker moment */
-    private _pickerMoment: T;
     @Input()
     get pickerMoment() {
         return this._pickerMoment;
@@ -107,11 +89,6 @@ export class OwlCalendarComponent<T>
     }
 
     @Input()
-    selectMode: SelectMode;
-
-    /** The currently selected moment. */
-    private _selected: T | null;
-    @Input()
     get selected(): T | null {
         return this._selected;
     }
@@ -121,7 +98,6 @@ export class OwlCalendarComponent<T>
         this._selected = this.getValidDate(value);
     }
 
-    private _selecteds: T[] = [];
     @Input()
     get selecteds(): T[] {
         return this._selecteds;
@@ -134,48 +110,12 @@ export class OwlCalendarComponent<T>
         });
     }
 
-    /**
-     * The view that the calendar should start in.
-     */
-    @Input()
-    startView: 'month' | 'year' | 'multi-years' = 'month';
-
-    /**
-     * Whether to hide dates in other months at the start or end of the current month.
-     * */
-    @Input()
-    hideOtherMonths: boolean;
-
-    /** Emits when the currently picker moment changes. */
-    @Output()
-    pickerMomentChange = new EventEmitter<T>();
-
-    /** Emits when the currently selected date changes. */
-    @Output()
-    selectedChange = new EventEmitter<T>();
-
-    /** Emits when any date is selected. */
-    @Output()
-    userSelection = new EventEmitter<void>();
-
-    /**
-     * Emits the selected year. This doesn't imply a change on the selected date
-     * */
-    @Output()
-    readonly yearSelected = new EventEmitter<T>();
-
-    /**
-     * Emits the selected month. This doesn't imply a change on the selected date
-     * */
-    @Output()
-    readonly monthSelected = new EventEmitter<T>();
-
     get periodButtonText(): string {
         return this.isMonthView
             ? this.dateTimeAdapter.format(
-                  this.pickerMoment,
-                  this.dateTimeFormats.monthYearLabel
-              )
+                this.pickerMoment,
+                this.dateTimeFormats.monthYearLabel
+            )
             : this.dateTimeAdapter.getYearName(this.pickerMoment);
     }
 
@@ -186,9 +126,9 @@ export class OwlCalendarComponent<T>
     }
 
     get prevButtonLabel(): string {
-        if (this._currentView === 'month') {
+        if (this._currentView === DateView.MONTH) {
             return this.pickerIntl.prevMonthLabel;
-        } else if (this._currentView === 'year') {
+        } else if (this._currentView === DateView.YEAR) {
             return this.pickerIntl.prevYearLabel;
         } else {
             return null;
@@ -196,21 +136,20 @@ export class OwlCalendarComponent<T>
     }
 
     get nextButtonLabel(): string {
-        if (this._currentView === 'month') {
+        if (this._currentView === DateView.MONTH) {
             return this.pickerIntl.nextMonthLabel;
-        } else if (this._currentView === 'year') {
+        } else if (this._currentView === DateView.YEAR) {
             return this.pickerIntl.nextYearLabel;
         } else {
             return null;
         }
     }
 
-    private _currentView: 'month' | 'year' | 'multi-years';
-    get currentView(): 'month' | 'year' | 'multi-years' {
+    get currentView(): DateViewType {
         return this._currentView;
     }
 
-    set currentView(view: 'month' | 'year' | 'multi-years') {
+    set currentView(view: DateViewType) {
         this._currentView = view;
         this.moveFocusOnNextTick = true;
     }
@@ -228,26 +167,12 @@ export class OwlCalendarComponent<T>
     }
 
     get showControlArrows(): boolean {
-        return this._currentView !== 'multi-years';
+        return this._currentView !== DateView.MULTI_YEARS;
     }
 
     get isMonthView() {
-        return this._currentView === 'month';
+        return this._currentView === DateView.MONTH;
     }
-
-    /**
-     * Date filter for the month and year view
-     */
-    public dateFilterForViews = (date: T) => {
-        return (
-            !!date &&
-            (!this.dateFilter || this.dateFilter(date)) &&
-            (!this.minDate ||
-                this.dateTimeAdapter.compare(date, this.minDate) >= 0) &&
-            (!this.maxDate ||
-                this.dateTimeAdapter.compare(date, this.maxDate) <= 0)
-        );
-    };
 
     /**
      * Bind class 'owl-dt-calendar' to host
@@ -255,15 +180,6 @@ export class OwlCalendarComponent<T>
     get owlDTCalendarClass(): boolean {
         return true;
     }
-
-    private intlChangesSub = Subscription.EMPTY;
-
-    /**
-     * Used for scheduling that focus should be moved to the active cell on the next tick.
-     * We need to schedule it, rather than do it immediately, because we have to wait
-     * for Angular to re-evaluate the view children.
-     */
-    private moveFocusOnNextTick = false;
 
     constructor(
         private elmRef: ElementRef,
@@ -280,7 +196,114 @@ export class OwlCalendarComponent<T>
         });
     }
 
-    public ngOnInit() {}
+    /**
+     * Date filter for the month and year view
+     * */
+    @Input()
+    dateFilter: (date: T) => boolean;
+
+    /**
+     * Set the first day of week
+     */
+    @Input()
+    firstDayOfWeek: number;
+
+    /** The minimum selectable date. */
+    private _minDate: T | null;
+
+    /** The maximum selectable date. */
+    private _maxDate: T | null;
+
+    /** The current picker moment */
+    private _pickerMoment: T;
+
+    @Input()
+    selectMode: SelectMode;
+
+    /** The currently selected moment. */
+    private _selected: T | null;
+
+    private _selecteds: T[] = [];
+
+    /**
+     * The view that the calendar should start in.
+     */
+    @Input()
+    startView: DateViewType = DateView.MONTH;
+
+    /**
+     * Whether to should only the year and multi-year views.
+     */
+    @Input()
+    yearOnly = false;
+
+    /**
+     * Whether to should only the multi-year view.
+     */
+    @Input()
+    multiyearOnly = false;
+
+    /**
+     * Whether to hide dates in other months at the start or end of the current month.
+     * */
+    @Input()
+    hideOtherMonths: boolean;
+
+    /** Emits when the currently picker moment changes. */
+    @Output()
+    pickerMomentChange = new EventEmitter<T>();
+
+    /** Emits when the selected date changes. */
+    @Output()
+    readonly dateClicked = new EventEmitter<T>();
+
+    /** Emits when the currently selected date changes. */
+    @Output()
+    readonly selectedChange = new EventEmitter<T>();
+
+    /** Emits when any date is selected. */
+    @Output()
+    readonly userSelection = new EventEmitter<void>();
+
+    /**
+     * Emits the selected year. This doesn't imply a change on the selected date
+     * */
+    @Output()
+    readonly yearSelected = new EventEmitter<T>();
+
+    /**
+     * Emits the selected month. This doesn't imply a change on the selected date
+     * */
+    @Output()
+    readonly monthSelected = new EventEmitter<T>();
+
+    private _currentView: DateViewType;
+
+    private intlChangesSub = Subscription.EMPTY;
+
+    /**
+     * Used for scheduling that focus should be moved to the active cell on the next tick.
+     * We need to schedule it, rather than do it immediately, because we have to wait
+     * for Angular to re-evaluate the view children.
+     */
+    private moveFocusOnNextTick = false;
+
+    /**
+     * Date filter for the month and year view
+     */
+    public dateFilterForViews = (date: T) => {
+        return (
+            !!date &&
+            (!this.dateFilter || this.dateFilter(date)) &&
+            (!this.minDate ||
+                this.dateTimeAdapter.compare(date, this.minDate) >= 0) &&
+            (!this.maxDate ||
+                this.dateTimeAdapter.compare(date, this.maxDate) <= 0)
+        );
+    };
+
+    public ngOnInit() {
+    }
 
     public ngAfterContentInit(): void {
         this._currentView = this.startView;
@@ -301,8 +324,19 @@ export class OwlCalendarComponent<T>
      * Toggle between month view and year view
      */
     public toggleViews(): void {
-        this.currentView =
-            this._currentView == 'month' ? 'multi-years' : 'month';
+        let nextView = null;
+        if (this._currentView === DateView.MONTH) {
+            nextView = DateView.MULTI_YEARS;
+        } else {
+            if (this.multiyearOnly) {
+                nextView = DateView.MULTI_YEARS;
+            } else if (this.yearOnly) {
+                nextView = this._currentView === DateView.YEAR ? DateView.MULTI_YEARS : DateView.YEAR;
+            } else {
+                nextView = DateView.MONTH;
+            }
+        }
+        this.currentView = nextView;
     }
 
     /**
@@ -332,6 +366,7 @@ export class OwlCalendarComponent<T>
             return;
         }
 
+        this.dateClicked.emit(date);
         this.selectedChange.emit(date);
 
         /*if ((this.isInSingleMode && !this.dateTimeAdapter.isSameDay(date, this.selected)) ||
@@ -345,10 +380,14 @@ export class OwlCalendarComponent<T>
      */
     public goToDateInView(
         date: T,
-        view: 'month' | 'year' | 'multi-years'
+        view: DateViewType
     ): void {
         this.handlePickerMomentChange(date);
-        this.currentView = view;
+        if ((!this.yearOnly && !this.multiyearOnly) ||
+            (this.multiyearOnly && (view !== DateView.MONTH && view !== DateView.YEAR)) ||
+            (this.yearOnly && view !== DateView.MONTH)) {
+            this.currentView = view;
+        }
         return;
     }
 
@@ -415,21 +454,21 @@ export class OwlCalendarComponent<T>
      * Whether the two dates represent the same view in the current view mode (month or year).
      */
     private isSameView(date1: T, date2: T): boolean {
-        if (this._currentView === 'month') {
+        if (this._currentView === DateView.MONTH) {
             return !!(
                 date1 &&
                 date2 &&
                 this.dateTimeAdapter.getYear(date1) ===
-                    this.dateTimeAdapter.getYear(date2) &&
+                this.dateTimeAdapter.getYear(date2) &&
                 this.dateTimeAdapter.getMonth(date1) ===
-                    this.dateTimeAdapter.getMonth(date2)
+                this.dateTimeAdapter.getMonth(date2)
             );
-        } else if (this._currentView === 'year') {
+        } else if (this._currentView === DateView.YEAR) {
             return !!(
                 date1 &&
                 date2 &&
                 this.dateTimeAdapter.getYear(date1) ===
-                    this.dateTimeAdapter.getYear(date2)
+                this.dateTimeAdapter.getYear(date2)
             );
         } else {
             return false;
@@ -441,7 +480,7 @@ export class OwlCalendarComponent<T>
      */
     private getValidDate(obj: any): T | null {
         return this.dateTimeAdapter.isDateInstance(obj) &&
-            this.dateTimeAdapter.isValid(obj)
+        this.dateTimeAdapter.isValid(obj)
             ? obj
             : null;
     }
