@@ -18,9 +18,8 @@ import {
     ViewChild
 } from '@angular/core';
 import { OwlDateTimeInputDirective } from './date-time-picker-input.directive';
-import { FormControl, FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { UntypedFormControl, FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { BrowserDynamicTestingModule } from '@angular/platform-browser-dynamic/testing';
 import { OwlDateTimeModule } from './date-time.module';
 import { OverlayContainer } from '@angular/cdk/overlay';
 import { OwlNativeDateTimeModule } from './adapter/native-date-time.module';
@@ -35,7 +34,8 @@ import { ENTER, ESCAPE, RIGHT_ARROW, UP_ARROW } from '@angular/cdk/keycodes';
 import { By } from '@angular/platform-browser';
 import { OwlDateTimeContainerComponent } from './date-time-picker-container.component';
 import { OwlDateTimeTriggerDirective } from './date-time-picker-trigger.directive';
-import { OWL_DATE_TIME_FORMATS } from './adapter/date-time-format.class';
+import { BrowserDynamicTestingModule } from '@angular/platform-browser-dynamic/testing';
+import {DateView} from './date-time.class';
 
 const JAN = 0,
     FEB = 1,
@@ -51,7 +51,7 @@ const JAN = 0,
     DEC = 11;
 
 describe('OwlDateTimeComponent', () => {
-    const SUPPORTS_INTL = typeof Intl != 'undefined';
+    const SUPPORTS_INTL = typeof Intl !== 'undefined';
 
     // Creates a test component fixture.
     function createComponent(
@@ -73,9 +73,7 @@ describe('OwlDateTimeComponent', () => {
         });
 
         TestBed.overrideModule(BrowserDynamicTestingModule, {
-            set: {
-                entryComponents: [entryComponents]
-            }
+            
         }).compileComponents();
 
         return TestBed.createComponent(component);
@@ -87,12 +85,12 @@ describe('OwlDateTimeComponent', () => {
 
     describe('with OwlNativeDateTimeModule', () => {
         describe('standard DateTimePicker', () => {
-            let fixture: ComponentFixture<StandardDateTimePicker>;
-            let testComponent: StandardDateTimePicker;
+            let fixture: ComponentFixture<StandardDateTimePickerComponent>;
+            let testComponent: StandardDateTimePickerComponent;
             let containerElement: HTMLElement;
 
             beforeEach(fakeAsync(() => {
-                fixture = createComponent(StandardDateTimePicker, [
+                fixture = createComponent(StandardDateTimePickerComponent, [
                     OwlNativeDateTimeModule
                 ]);
                 fixture.detectChanges();
@@ -200,19 +198,18 @@ describe('OwlDateTimeComponent', () => {
                 fixture.detectChanges();
                 flush();
 
-                const popup = document.querySelector('.cdk-overlay-pane')!;
+                const popup = document.querySelector('.cdk-overlay-pane');
                 expect(popup).not.toBeNull();
                 expect(
-                    parseInt(getComputedStyle(popup).height as string)
+                    parseInt(getComputedStyle(popup).height, 10)
                 ).not.toBe(0);
 
                 testComponent.dateTimePicker.close();
                 fixture.detectChanges();
                 flush();
 
-                expect(parseInt(getComputedStyle(popup).height as string)).toBe(
-                    0
-                );
+                const newHeight = getComputedStyle(popup).height;
+                expect(newHeight).toBe('');
             }));
 
             it('should close the popup when pressing ESCAPE', fakeAsync(() => {
@@ -265,12 +262,12 @@ describe('OwlDateTimeComponent', () => {
                     'Expected dateTimePicker to be opened.'
                 );
 
-                let containerDebugElement = fixture.debugElement.query(
+                const containerDebugElement = fixture.debugElement.query(
                     By.directive(OwlDateTimeContainerComponent)
                 );
                 containerElement = containerDebugElement.nativeElement;
 
-                let btns = containerElement.querySelectorAll(
+                const btns = containerElement.querySelectorAll(
                     '.owl-dt-container-control-button'
                 );
                 dispatchMouseEvent(btns[0], 'click'); // 'Cancel' button
@@ -292,18 +289,18 @@ describe('OwlDateTimeComponent', () => {
                     'Expected dateTimePicker to be opened.'
                 );
 
-                let containerDebugElement = fixture.debugElement.query(
+                const containerDebugElement = fixture.debugElement.query(
                     By.directive(OwlDateTimeContainerComponent)
                 );
                 containerElement = containerDebugElement.nativeElement;
 
-                let monthCell = containerElement.querySelector(
+                const monthCell = containerElement.querySelector(
                     '[aria-label="January 2, 2020"]'
                 );
                 (monthCell as HTMLElement).click();
                 fixture.detectChanges();
 
-                let btns = containerElement.querySelectorAll(
+                const btns = containerElement.querySelectorAll(
                     '.owl-dt-container-control-button'
                 );
                 dispatchMouseEvent(btns[0], 'click'); // 'Cancel' button
@@ -328,7 +325,7 @@ describe('OwlDateTimeComponent', () => {
                     'Expected dateTimePicker to be opened.'
                 );
 
-                let containerDebugElement = fixture.debugElement.query(
+                const containerDebugElement = fixture.debugElement.query(
                     By.directive(OwlDateTimeContainerComponent)
                 );
                 containerElement = containerDebugElement.nativeElement;
@@ -336,7 +333,7 @@ describe('OwlDateTimeComponent', () => {
                     containerDebugElement.componentInstance.pickerMoment
                 ).toEqual(new Date(2020, JAN, 1));
 
-                let btns = containerElement.querySelectorAll(
+                const btns = containerElement.querySelectorAll(
                     '.owl-dt-container-control-button'
                 );
                 dispatchMouseEvent(btns[1], 'click'); // 'Set' button
@@ -361,7 +358,7 @@ describe('OwlDateTimeComponent', () => {
                     'Expected dateTimePicker to be opened.'
                 );
 
-                let containerDebugElement = fixture.debugElement.query(
+                const containerDebugElement = fixture.debugElement.query(
                     By.directive(OwlDateTimeContainerComponent)
                 );
                 containerElement = containerDebugElement.nativeElement;
@@ -369,13 +366,13 @@ describe('OwlDateTimeComponent', () => {
                     containerDebugElement.componentInstance.pickerMoment
                 ).toEqual(new Date(2020, JAN, 1));
 
-                let monthCell = containerElement.querySelector(
+                const monthCell = containerElement.querySelector(
                     '[aria-label="January 2, 2020"]'
                 );
                 (monthCell as HTMLElement).click();
                 fixture.detectChanges();
 
-                let btns = containerElement.querySelectorAll(
+                const btns = containerElement.querySelectorAll(
                     '.owl-dt-container-control-button'
                 );
                 dispatchMouseEvent(btns[1], 'click'); // 'Set' button
@@ -398,7 +395,7 @@ describe('OwlDateTimeComponent', () => {
             });
 
             it('input should aria-owns owl-date-time-container after opened in popup mode', fakeAsync(() => {
-                let inputEl = fixture.debugElement.query(By.css('input'))
+                const inputEl = fixture.debugElement.query(By.css('input'))
                     .nativeElement;
                 expect(inputEl.getAttribute('aria-owns')).toBeNull();
 
@@ -407,10 +404,10 @@ describe('OwlDateTimeComponent', () => {
                 flush();
                 fixture.detectChanges();
 
-                let ownedElementId = inputEl.getAttribute('aria-owns');
+                const ownedElementId = inputEl.getAttribute('aria-owns');
                 expect(ownedElementId).not.toBeNull();
 
-                let ownedElement = document.getElementById(ownedElementId);
+                const ownedElement = document.getElementById(ownedElementId);
                 expect(ownedElement).not.toBeNull();
                 expect((ownedElement as Element).tagName.toLowerCase()).toBe(
                     'owl-date-time-container'
@@ -421,7 +418,7 @@ describe('OwlDateTimeComponent', () => {
                 testComponent.pickerMode = 'dialog';
                 fixture.detectChanges();
 
-                let inputEl = fixture.debugElement.query(By.css('input'))
+                const inputEl = fixture.debugElement.query(By.css('input'))
                     .nativeElement;
                 expect(inputEl.getAttribute('aria-owns')).toBeNull();
 
@@ -430,10 +427,10 @@ describe('OwlDateTimeComponent', () => {
                 flush();
                 fixture.detectChanges();
 
-                let ownedElementId = inputEl.getAttribute('aria-owns');
+                const ownedElementId = inputEl.getAttribute('aria-owns');
                 expect(ownedElementId).not.toBeNull();
 
-                let ownedElement = document.getElementById(ownedElementId);
+                const ownedElement = document.getElementById(ownedElementId);
                 expect(ownedElement).not.toBeNull();
                 expect((ownedElement as Element).tagName.toLowerCase()).toBe(
                     'owl-date-time-container'
@@ -480,11 +477,11 @@ describe('OwlDateTimeComponent', () => {
                         'Expected dateTimePicker to be opened.'
                     );
 
-                    let containerDebugElement = fixture.debugElement.query(
+                    const containerDebugElement = fixture.debugElement.query(
                         By.directive(OwlDateTimeContainerComponent)
                     );
                     containerElement = containerDebugElement.nativeElement;
-                    let btns = containerElement.querySelectorAll(
+                    const btns = containerElement.querySelectorAll(
                         '.owl-dt-container-control-button'
                     );
 
@@ -500,7 +497,7 @@ describe('OwlDateTimeComponent', () => {
                         'Expected dateTimePicker to be opened.'
                     );
 
-                    let containerDebugElement = fixture.debugElement.query(
+                    const containerDebugElement = fixture.debugElement.query(
                         By.directive(OwlDateTimeContainerComponent)
                     );
                     containerElement = containerDebugElement.nativeElement;
@@ -508,7 +505,7 @@ describe('OwlDateTimeComponent', () => {
                         containerDebugElement.componentInstance.pickerMoment
                     ).toEqual(new Date(2020, JAN, 1));
 
-                    let dateCell = containerElement.querySelector(
+                    const dateCell = containerElement.querySelector(
                         '[aria-label="January 2, 2020"]'
                     );
                     dispatchMouseEvent(dateCell, 'click');
@@ -524,46 +521,47 @@ describe('OwlDateTimeComponent', () => {
                     );
                 }));
 
-                it('should update input value to clicked date value and close popup panel when date cell is clicked via pressing enter', fakeAsync(() => {
-                    testComponent.dateTimePicker.open();
-                    fixture.detectChanges();
-                    flush();
-                    expect(testComponent.dateTimePicker.opened).toBe(
-                        true,
-                        'Expected dateTimePicker to be opened.'
-                    );
+                it('should update input value to clicked date value and close popup panel when date cell is clicked via pressing enter',
+                    fakeAsync(() => {
+                        testComponent.dateTimePicker.open();
+                        fixture.detectChanges();
+                        flush();
+                        expect(testComponent.dateTimePicker.opened).toBe(
+                            true,
+                            'Expected dateTimePicker to be opened.'
+                        );
 
-                    let containerDebugElement = fixture.debugElement.query(
-                        By.directive(OwlDateTimeContainerComponent)
-                    );
-                    containerElement = containerDebugElement.nativeElement;
-                    expect(
-                        containerDebugElement.componentInstance.pickerMoment
-                    ).toEqual(new Date(2020, JAN, 1));
+                        const containerDebugElement = fixture.debugElement.query(
+                            By.directive(OwlDateTimeContainerComponent)
+                        );
+                        containerElement = containerDebugElement.nativeElement;
+                        expect(
+                            containerDebugElement.componentInstance.pickerMoment
+                        ).toEqual(new Date(2020, JAN, 1));
 
-                    let calendarBodyEl = containerElement.querySelector(
-                        '.owl-dt-calendar-body'
-                    );
+                        const calendarBodyEl = containerElement.querySelector(
+                            '.owl-dt-calendar-body'
+                        );
 
-                    dispatchKeyboardEvent(
-                        calendarBodyEl,
-                        'keydown',
-                        RIGHT_ARROW
-                    );
-                    fixture.detectChanges();
-                    flush();
-                    dispatchKeyboardEvent(calendarBodyEl, 'keydown', ENTER);
-                    fixture.detectChanges();
-                    flush();
+                        dispatchKeyboardEvent(
+                            calendarBodyEl,
+                            'keydown',
+                            RIGHT_ARROW
+                        );
+                        fixture.detectChanges();
+                        flush();
+                        dispatchKeyboardEvent(calendarBodyEl, 'keydown', ENTER);
+                        fixture.detectChanges();
+                        flush();
 
-                    expect(testComponent.dateTimePicker.opened).toBe(
-                        false,
-                        'Expected dateTimePicker to be closed.'
-                    );
-                    expect(testComponent.dateTimePickerInput.value).toEqual(
-                        new Date(2020, JAN, 2)
-                    );
-                }));
+                        expect(testComponent.dateTimePicker.opened).toBe(
+                            false,
+                            'Expected dateTimePicker to be closed.'
+                        );
+                        expect(testComponent.dateTimePickerInput.value).toEqual(
+                            new Date(2020, JAN, 2)
+                        );
+                    }));
 
                 it('should close popup panel when click on the selected date', fakeAsync(() => {
                     testComponent.dateTimePicker.open();
@@ -574,7 +572,7 @@ describe('OwlDateTimeComponent', () => {
                         'Expected dateTimePicker to be opened.'
                     );
 
-                    let containerDebugElement = fixture.debugElement.query(
+                    const containerDebugElement = fixture.debugElement.query(
                         By.directive(OwlDateTimeContainerComponent)
                     );
                     containerElement = containerDebugElement.nativeElement;
@@ -585,7 +583,7 @@ describe('OwlDateTimeComponent', () => {
                         new Date(2020, JAN, 1)
                     );
 
-                    let dateCell = containerElement.querySelector(
+                    const dateCell = containerElement.querySelector(
                         '[aria-label="January 1, 2020"]'
                     );
                     dispatchMouseEvent(dateCell, 'click');
@@ -625,11 +623,11 @@ describe('OwlDateTimeComponent', () => {
                         'Expected dateTimePicker to be opened.'
                     );
 
-                    let containerDebugElement = fixture.debugElement.query(
+                    const containerDebugElement = fixture.debugElement.query(
                         By.directive(OwlDateTimeContainerComponent)
                     );
                     containerElement = containerDebugElement.nativeElement;
-                    let btns = containerElement.querySelectorAll(
+                    const btns = containerElement.querySelectorAll(
                         '.owl-dt-container-control-button'
                     );
 
@@ -639,12 +637,12 @@ describe('OwlDateTimeComponent', () => {
         });
 
         describe('range DateTimePicker', () => {
-            let fixture: ComponentFixture<RangeDateTimePicker>;
-            let testComponent: RangeDateTimePicker;
+            let fixture: ComponentFixture<RangeDateTimePickerComponent>;
+            let testComponent: RangeDateTimePickerComponent;
             let containerElement: HTMLElement;
 
             beforeEach(fakeAsync(() => {
-                fixture = createComponent(RangeDateTimePicker, [
+                fixture = createComponent(RangeDateTimePickerComponent, [
                     OwlNativeDateTimeModule
                 ]);
                 fixture.detectChanges();
@@ -661,7 +659,7 @@ describe('OwlDateTimeComponent', () => {
                 if (SUPPORTS_INTL) {
                     expect(
                         fixture.nativeElement.querySelector('input').value
-                    ).toBe('1/1/2020, 12:00 AM ~ 2/1/2020, 12:00 AM');
+                    ).toBe('1/1/2020, 12:00 AM - 2/1/2020, 12:00 AM');
                 }
             });
 
@@ -669,7 +667,7 @@ describe('OwlDateTimeComponent', () => {
                 testComponent.dateTimePicker.open();
                 fixture.detectChanges();
 
-                let containerDebugElement = fixture.debugElement.query(
+                const containerDebugElement = fixture.debugElement.query(
                     By.directive(OwlDateTimeContainerComponent)
                 );
                 expect(
@@ -687,12 +685,12 @@ describe('OwlDateTimeComponent', () => {
 
                 expect(testComponent.dateTimePicker.selecteds.length).toBe(0);
 
-                let containerDebugElement = fixture.debugElement.query(
+                const containerDebugElement = fixture.debugElement.query(
                     By.directive(OwlDateTimeContainerComponent)
                 );
                 containerElement = containerDebugElement.nativeElement;
 
-                let dateCell = containerElement.querySelector(
+                const dateCell = containerElement.querySelector(
                     '[aria-label="January 2, 2020"]'
                 );
                 dispatchMouseEvent(dateCell, 'click');
@@ -716,12 +714,12 @@ describe('OwlDateTimeComponent', () => {
 
                 expect(testComponent.dateTimePicker.selecteds.length).toBe(2);
 
-                let containerDebugElement = fixture.debugElement.query(
+                const containerDebugElement = fixture.debugElement.query(
                     By.directive(OwlDateTimeContainerComponent)
                 );
                 containerElement = containerDebugElement.nativeElement;
 
-                let dateCell = containerElement.querySelector(
+                const dateCell = containerElement.querySelector(
                     '[aria-label="January 2, 2020"]'
                 );
                 dispatchMouseEvent(dateCell, 'click');
@@ -738,37 +736,38 @@ describe('OwlDateTimeComponent', () => {
                 expect(testComponent.dateTimePicker.selecteds[1]).toBe(null);
             }));
 
-            it('clicking the dateCell should set the rangeFrom value when dateCell value is before the old rangeFrom value', fakeAsync(() => {
-                testComponent.dates = [new Date(2020, JAN, 2), null];
-                fixture.detectChanges();
+            it('clicking the dateCell should set the rangeFrom value when dateCell value is before the old rangeFrom value',
+                fakeAsync(() => {
+                    testComponent.dates = [new Date(2020, JAN, 2), null];
+                    fixture.detectChanges();
 
-                testComponent.dateTimePicker.open();
-                fixture.detectChanges();
-                flush();
+                    testComponent.dateTimePicker.open();
+                    fixture.detectChanges();
+                    flush();
 
-                expect(testComponent.dateTimePicker.selecteds.length).toBe(2);
+                    expect(testComponent.dateTimePicker.selecteds.length).toBe(2);
 
-                let containerDebugElement = fixture.debugElement.query(
-                    By.directive(OwlDateTimeContainerComponent)
-                );
-                containerElement = containerDebugElement.nativeElement;
+                    const containerDebugElement = fixture.debugElement.query(
+                        By.directive(OwlDateTimeContainerComponent)
+                    );
+                    containerElement = containerDebugElement.nativeElement;
 
-                let dateCell = containerElement.querySelector(
-                    '[aria-label="January 1, 2020"]'
-                );
-                dispatchMouseEvent(dateCell, 'click');
-                fixture.detectChanges();
-                flush();
+                    const dateCell = containerElement.querySelector(
+                        '[aria-label="January 1, 2020"]'
+                    );
+                    dispatchMouseEvent(dateCell, 'click');
+                    fixture.detectChanges();
+                    flush();
 
-                expect(
-                    containerDebugElement.componentInstance.activeSelectedIndex
-                ).toBe(0);
-                expect(testComponent.dateTimePicker.selecteds.length).toBe(2);
-                expect(testComponent.dateTimePicker.selecteds[0]).toEqual(
-                    new Date(2020, JAN, 1)
-                );
-                expect(testComponent.dateTimePicker.selecteds[1]).toBe(null);
-            }));
+                    expect(
+                        containerDebugElement.componentInstance.activeSelectedIndex
+                    ).toBe(0);
+                    expect(testComponent.dateTimePicker.selecteds.length).toBe(2);
+                    expect(testComponent.dateTimePicker.selecteds[0]).toEqual(
+                        new Date(2020, JAN, 1)
+                    );
+                    expect(testComponent.dateTimePicker.selecteds[1]).toBe(null);
+                }));
 
             it('clicking the dateCell should set the rangeTo value when rangeFrom already had value', fakeAsync(() => {
                 testComponent.dates = [new Date(2020, JAN, 2), null];
@@ -780,12 +779,12 @@ describe('OwlDateTimeComponent', () => {
 
                 expect(testComponent.dateTimePicker.selecteds.length).toBe(2);
 
-                let containerDebugElement = fixture.debugElement.query(
+                const containerDebugElement = fixture.debugElement.query(
                     By.directive(OwlDateTimeContainerComponent)
                 );
                 containerElement = containerDebugElement.nativeElement;
 
-                let dateCell = containerElement.querySelector(
+                const dateCell = containerElement.querySelector(
                     '[aria-label="January 3, 2020"]'
                 );
                 dispatchMouseEvent(dateCell, 'click');
@@ -804,16 +803,82 @@ describe('OwlDateTimeComponent', () => {
                 );
             }));
 
+            it('if startAt value is set, the start time value should be shown in the rangeTo calendar start time', fakeAsync(() => {
+                testComponent.startAt = new Date('1/19/2020, 09:33 AM');
+                testComponent.dateTimePicker.open();
+                fixture.detectChanges();
+                flush();
+
+                const containerDebugElement = fixture.debugElement.query(
+                    By.directive(OwlDateTimeContainerComponent)
+                );
+                containerElement = containerDebugElement.nativeElement;
+                const timeCells = containerElement.querySelectorAll<HTMLInputElement>('.owl-dt-timer-input');
+
+                expect(timeCells[0].value).toEqual('09');
+                expect(timeCells[1].value).toEqual('33');
+            }));
+
+
+            it('if endAt value is set, the end time value should be shown in the rangeTo calendar end time', fakeAsync(() => {
+                testComponent.dates = [new Date(2020, JAN, 2), null];
+                testComponent.endAt = new Date('1/19/2020, 10:55 PM');
+                fixture.detectChanges();
+
+                testComponent.dateTimePicker.open();
+                fixture.detectChanges();
+                flush();
+
+                const containerDebugElement = fixture.debugElement.query(
+                    By.directive(OwlDateTimeContainerComponent)
+                );
+                containerElement = containerDebugElement.nativeElement;
+                const dateCell = containerElement.querySelector(
+                    '[aria-label="January 20, 2020"]'
+                );
+                dispatchMouseEvent(dateCell, 'click');
+                fixture.detectChanges();
+                flush();
+
+                const timeCells = containerElement.querySelectorAll<HTMLInputElement>('.owl-dt-timer-input');
+
+                expect(timeCells[0].value).toEqual('22');
+                expect(timeCells[1].value).toEqual('55');
+            }));
+
+            it('auto select previous time if dates is selected', fakeAsync(() => {
+                testComponent.dates = [new Date('1/19/2020, 09:33 AM'), new Date('1/22/2020, 10:44 PM')];
+                testComponent.dateTimePicker.open();
+                fixture.detectChanges();
+
+                const containerDebugElement = fixture.debugElement.query(
+                    By.directive(OwlDateTimeContainerComponent)
+                );
+                containerElement = containerDebugElement.nativeElement;
+                const dateCell = containerElement.querySelector(
+                    '[aria-label="January 20, 2020"]'
+                );
+                dispatchMouseEvent(dateCell, 'click');
+                fixture.detectChanges();
+                flush();
+
+                const timeInput = containerElement.querySelectorAll<HTMLInputElement>('.owl-dt-timer-input');
+
+                expect(timeInput[0].value).toEqual('09');
+                expect(timeInput[1].value).toEqual('33');
+
+            }));
+
             it('should have the container info row', () => {
                 testComponent.dateTimePicker.open();
                 fixture.detectChanges();
 
-                let containerDebugElement = fixture.debugElement.query(
+                const containerDebugElement = fixture.debugElement.query(
                     By.directive(OwlDateTimeContainerComponent)
                 );
                 containerElement = containerDebugElement.nativeElement;
 
-                let infoRow = containerElement.querySelector(
+                const infoRow = containerElement.querySelector(
                     '.owl-dt-container-info'
                 );
 
@@ -824,12 +889,12 @@ describe('OwlDateTimeComponent', () => {
                 testComponent.dateTimePicker.open();
                 fixture.detectChanges();
 
-                let containerDebugElement = fixture.debugElement.query(
+                const containerDebugElement = fixture.debugElement.query(
                     By.directive(OwlDateTimeContainerComponent)
                 );
                 containerElement = containerDebugElement.nativeElement;
 
-                let radioBtns = containerElement.querySelectorAll(
+                const radioBtns = containerElement.querySelectorAll(
                     '.owl-dt-container-range'
                 );
 
@@ -855,7 +920,7 @@ describe('OwlDateTimeComponent', () => {
                     if (SUPPORTS_INTL) {
                         expect(
                             fixture.nativeElement.querySelector('input').value
-                        ).toBe('1/1/2020 ~ 2/1/2020');
+                        ).toBe('1/1/2020 - 2/1/2020');
                     }
                 });
 
@@ -871,12 +936,12 @@ describe('OwlDateTimeComponent', () => {
                         0
                     );
 
-                    let containerDebugElement = fixture.debugElement.query(
+                    const containerDebugElement = fixture.debugElement.query(
                         By.directive(OwlDateTimeContainerComponent)
                     );
                     containerElement = containerDebugElement.nativeElement;
 
-                    let dateCell = containerElement.querySelector(
+                    const dateCell = containerElement.querySelector(
                         '[aria-label="January 2, 2020"]'
                     );
                     dispatchMouseEvent(dateCell, 'click');
@@ -889,51 +954,52 @@ describe('OwlDateTimeComponent', () => {
                     );
                 }));
 
-                it('should close the dateTimePicker popup panel when both the rangeFrom and the rangeTo value are selected', fakeAsync(() => {
-                    testComponent.dates = [];
-                    fixture.detectChanges();
+                it('should close the dateTimePicker popup panel when both the rangeFrom and the rangeTo value are selected',
+                    fakeAsync(() => {
+                        testComponent.dates = [];
+                        fixture.detectChanges();
 
-                    testComponent.dateTimePicker.open();
-                    fixture.detectChanges();
-                    flush();
+                        testComponent.dateTimePicker.open();
+                        fixture.detectChanges();
+                        flush();
 
-                    expect(testComponent.dateTimePicker.selecteds.length).toBe(
-                        0
-                    );
+                        expect(testComponent.dateTimePicker.selecteds.length).toBe(
+                            0
+                        );
 
-                    let containerDebugElement = fixture.debugElement.query(
-                        By.directive(OwlDateTimeContainerComponent)
-                    );
-                    containerElement = containerDebugElement.nativeElement;
+                        const containerDebugElement = fixture.debugElement.query(
+                            By.directive(OwlDateTimeContainerComponent)
+                        );
+                        containerElement = containerDebugElement.nativeElement;
 
-                    let dateCell = containerElement.querySelector(
-                        '[aria-label="January 2, 2020"]'
-                    );
-                    dispatchMouseEvent(dateCell, 'click');
-                    fixture.detectChanges();
-                    flush();
+                        let dateCell = containerElement.querySelector(
+                            '[aria-label="January 2, 2020"]'
+                        );
+                        dispatchMouseEvent(dateCell, 'click');
+                        fixture.detectChanges();
+                        flush();
 
-                    dateCell = containerElement.querySelector(
-                        '[aria-label="January 3, 2020"]'
-                    );
-                    dispatchMouseEvent(dateCell, 'click');
-                    fixture.detectChanges();
-                    flush();
+                        dateCell = containerElement.querySelector(
+                            '[aria-label="January 3, 2020"]'
+                        );
+                        dispatchMouseEvent(dateCell, 'click');
+                        fixture.detectChanges();
+                        flush();
 
-                    expect(testComponent.dateTimePicker.selecteds.length).toBe(
-                        2
-                    );
-                    expect(testComponent.dateTimePicker.opened).toBe(
-                        false,
-                        'Expected dateTimePicker to be closed.'
-                    );
-                }));
+                        expect(testComponent.dateTimePicker.selecteds.length).toBe(
+                            2
+                        );
+                        expect(testComponent.dateTimePicker.opened).toBe(
+                            false,
+                            'Expected dateTimePicker to be closed.'
+                        );
+                    }));
             });
         });
 
         describe('DateTimePicker with too many inputs', () => {
             it('should throw when multiple inputs registered', fakeAsync(() => {
-                let fixture = createComponent(MultiInputDateTimePicker, [
+                const fixture = createComponent(MultiInputDateTimePickerComponent, [
                     OwlNativeDateTimeModule
                 ]);
                 expect(() => fixture.detectChanges()).toThrow();
@@ -941,11 +1007,11 @@ describe('OwlDateTimeComponent', () => {
         });
 
         describe('DateTimePicker with no input', () => {
-            let fixture: ComponentFixture<NoInputDateTimePicker>;
-            let testComponent: NoInputDateTimePicker;
+            let fixture: ComponentFixture<NoInputDateTimePickerComponent>;
+            let testComponent: NoInputDateTimePickerComponent;
 
             beforeEach(fakeAsync(() => {
-                fixture = createComponent(NoInputDateTimePicker, [
+                fixture = createComponent(NoInputDateTimePickerComponent, [
                     OwlNativeDateTimeModule
                 ]);
                 fixture.detectChanges();
@@ -970,11 +1036,11 @@ describe('OwlDateTimeComponent', () => {
         });
 
         describe('DateTimePicker with startAt', () => {
-            let fixture: ComponentFixture<DateTimePickerWithStartAt>;
-            let testComponent: DateTimePickerWithStartAt;
+            let fixture: ComponentFixture<DateTimePickerWithStartAtComponent>;
+            let testComponent: DateTimePickerWithStartAtComponent;
 
             beforeEach(fakeAsync(() => {
-                fixture = createComponent(DateTimePickerWithStartAt, [
+                fixture = createComponent(DateTimePickerWithStartAtComponent, [
                     OwlNativeDateTimeModule
                 ]);
                 fixture.detectChanges();
@@ -995,13 +1061,13 @@ describe('OwlDateTimeComponent', () => {
         });
 
         describe('DateTimePicker with startView', () => {
-            let fixture: ComponentFixture<DateTimePickerWithStartView>;
-            let testComponent: DateTimePickerWithStartView;
+            let fixture: ComponentFixture<DateTimePickerWithStartViewComponent>;
+            let testComponent: DateTimePickerWithStartViewComponent;
             let containerDebugElement;
             let containerElement;
 
             beforeEach(() => {
-                fixture = createComponent(DateTimePickerWithStartView, [
+                fixture = createComponent(DateTimePickerWithStartViewComponent, [
                     OwlNativeDateTimeModule
                 ]);
                 fixture.detectChanges();
@@ -1016,7 +1082,7 @@ describe('OwlDateTimeComponent', () => {
 
             describe('set to year', () => {
                 beforeEach(() => {
-                    testComponent.startView = 'year';
+                    testComponent.startView = DateView.YEAR;
                     fixture.detectChanges();
                 });
 
@@ -1028,7 +1094,7 @@ describe('OwlDateTimeComponent', () => {
                         By.directive(OwlDateTimeContainerComponent)
                     );
                     containerElement = containerDebugElement.nativeElement;
-                    let yearTable = containerElement.querySelector(
+                    const yearTable = containerElement.querySelector(
                         'table.owl-dt-calendar-year-table'
                     );
                     expect(yearTable).toBeTruthy();
@@ -1061,7 +1127,7 @@ describe('OwlDateTimeComponent', () => {
 
             describe('set to multi-years', () => {
                 beforeEach(() => {
-                    testComponent.startView = 'multi-years';
+                    testComponent.startView = DateView.MULTI_YEARS;
                     fixture.detectChanges();
                 });
 
@@ -1073,7 +1139,7 @@ describe('OwlDateTimeComponent', () => {
                         By.directive(OwlDateTimeContainerComponent)
                     );
                     containerElement = containerDebugElement.nativeElement;
-                    let multiYearTable = containerElement.querySelector(
+                    const multiYearTable = containerElement.querySelector(
                         'table.owl-dt-calendar-multi-year-table'
                     );
                     expect(multiYearTable).toBeTruthy();
@@ -1106,11 +1172,11 @@ describe('OwlDateTimeComponent', () => {
         });
 
         describe('DateTimePicker with NgModel', () => {
-            let fixture: ComponentFixture<DateTimePickerWithNgModel>;
-            let testComponent: DateTimePickerWithNgModel;
+            let fixture: ComponentFixture<DateTimePickerWithNgModelComponent>;
+            let testComponent: DateTimePickerWithNgModelComponent;
 
             beforeEach(fakeAsync(() => {
-                fixture = createComponent(DateTimePickerWithNgModel, [
+                fixture = createComponent(DateTimePickerWithNgModelComponent, [
                     OwlNativeDateTimeModule
                 ]);
                 fixture.detectChanges();
@@ -1131,7 +1197,7 @@ describe('OwlDateTimeComponent', () => {
                 expect(testComponent.dateTimePickerInput.value).toBeNull();
                 expect(testComponent.dateTimePicker.selected).toBeNull();
 
-                let selected = new Date(2017, JAN, 1);
+                const selected = new Date(2017, JAN, 1);
                 testComponent.moment = selected;
                 fixture.detectChanges();
                 flush();
@@ -1147,7 +1213,7 @@ describe('OwlDateTimeComponent', () => {
                 expect(testComponent.moment).toBeNull();
                 expect(testComponent.dateTimePickerInput.value).toBeNull();
 
-                let selected = new Date(2017, JAN, 1);
+                const selected = new Date(2017, JAN, 1);
                 testComponent.dateTimePicker.select(selected);
                 fixture.detectChanges();
                 flush();
@@ -1163,7 +1229,7 @@ describe('OwlDateTimeComponent', () => {
             }));
 
             it('should mark input dirty after input event', () => {
-                let inputEl = fixture.debugElement.query(By.css('input'))
+                const inputEl = fixture.debugElement.query(By.css('input'))
                     .nativeElement;
 
                 expect(inputEl.classList).toContain('ng-pristine');
@@ -1176,7 +1242,7 @@ describe('OwlDateTimeComponent', () => {
             });
 
             it('should mark input dirty after date selected', fakeAsync(() => {
-                let inputEl = fixture.debugElement.query(By.css('input'))
+                const inputEl = fixture.debugElement.query(By.css('input'))
                     .nativeElement;
 
                 expect(inputEl.classList).toContain('ng-pristine');
@@ -1193,7 +1259,7 @@ describe('OwlDateTimeComponent', () => {
             }));
 
             it('should not mark dirty after model change', fakeAsync(() => {
-                let inputEl = fixture.debugElement.query(By.css('input'))
+                const inputEl = fixture.debugElement.query(By.css('input'))
                     .nativeElement;
 
                 expect(inputEl.classList).toContain('ng-pristine');
@@ -1207,7 +1273,7 @@ describe('OwlDateTimeComponent', () => {
             }));
 
             it('should mark input touched on blur', () => {
-                let inputEl = fixture.debugElement.query(By.css('input'))
+                const inputEl = fixture.debugElement.query(By.css('input'))
                     .nativeElement;
 
                 expect(inputEl.classList).toContain('ng-untouched');
@@ -1238,7 +1304,7 @@ describe('OwlDateTimeComponent', () => {
             });
 
             it('should mark input touched on calendar selection', fakeAsync(() => {
-                let inputEl = fixture.debugElement.query(By.css('input'))
+                const inputEl = fixture.debugElement.query(By.css('input'))
                     .nativeElement;
 
                 expect(inputEl.classList).toContain('ng-untouched');
@@ -1271,8 +1337,8 @@ describe('OwlDateTimeComponent', () => {
                         0
                     );
 
-                    let from = new Date(2017, JAN, 1);
-                    let to = new Date(2017, JAN, 3);
+                    const from = new Date(2017, JAN, 1);
+                    const to = new Date(2017, JAN, 3);
                     testComponent.moment = [from, to];
                     fixture.detectChanges();
                     flush();
@@ -1304,8 +1370,8 @@ describe('OwlDateTimeComponent', () => {
                         testComponent.dateTimePickerInput.values.length
                     ).toBe(0);
 
-                    let from = new Date(2017, JAN, 1);
-                    let to = new Date(2017, JAN, 3);
+                    const from = new Date(2017, JAN, 1);
+                    const to = new Date(2017, JAN, 3);
                     testComponent.dateTimePicker.select([from, to]);
                     fixture.detectChanges();
                     flush();
@@ -1345,7 +1411,7 @@ describe('OwlDateTimeComponent', () => {
                         0
                     );
 
-                    let from = new Date(2017, JAN, 1);
+                    const from = new Date(2017, JAN, 1);
                     testComponent.moment = [from];
                     fixture.detectChanges();
                     flush();
@@ -1366,8 +1432,8 @@ describe('OwlDateTimeComponent', () => {
                 }));
 
                 it('should only update fromValue when date is selected', fakeAsync(() => {
-                    let from = new Date(2017, JAN, 1);
-                    let to = new Date(2017, JAN, 3);
+                    const from = new Date(2017, JAN, 1);
+                    const to = new Date(2017, JAN, 3);
                     testComponent.moment = [from, to];
                     fixture.detectChanges();
                     flush();
@@ -1377,8 +1443,8 @@ describe('OwlDateTimeComponent', () => {
                     fixture.detectChanges();
                     flush();
 
-                    let newSelectedFrom = new Date(2017, JAN, 2);
-                    let containerDebugElement = fixture.debugElement.query(
+                    const newSelectedFrom = new Date(2017, JAN, 2);
+                    const containerDebugElement = fixture.debugElement.query(
                         By.directive(OwlDateTimeContainerComponent)
                     );
                     containerDebugElement.componentInstance.dateSelected(
@@ -1408,8 +1474,8 @@ describe('OwlDateTimeComponent', () => {
                 }));
 
                 it('should update fromValue and set toValue to null when date is selected after toValue', fakeAsync(() => {
-                    let from = new Date(2017, JAN, 1);
-                    let to = new Date(2017, JAN, 3);
+                    const from = new Date(2017, JAN, 1);
+                    const to = new Date(2017, JAN, 3);
                     testComponent.moment = [from, to];
                     fixture.detectChanges();
                     flush();
@@ -1419,8 +1485,8 @@ describe('OwlDateTimeComponent', () => {
                     fixture.detectChanges();
                     flush();
 
-                    let newSelectedFrom = new Date(2017, JAN, 4);
-                    let containerDebugElement = fixture.debugElement.query(
+                    const newSelectedFrom = new Date(2017, JAN, 4);
+                    const containerDebugElement = fixture.debugElement.query(
                         By.directive(OwlDateTimeContainerComponent)
                     );
                     containerDebugElement.componentInstance.dateSelected(
@@ -1467,7 +1533,7 @@ describe('OwlDateTimeComponent', () => {
                         0
                     );
 
-                    let to = new Date(2017, JAN, 3);
+                    const to = new Date(2017, JAN, 3);
                     testComponent.moment = [null, to];
                     fixture.detectChanges();
                     flush();
@@ -1488,8 +1554,8 @@ describe('OwlDateTimeComponent', () => {
                 }));
 
                 it('should only update toValue when date is selected', fakeAsync(() => {
-                    let from = new Date(2017, JAN, 1);
-                    let to = new Date(2017, JAN, 3);
+                    const from = new Date(2017, JAN, 1);
+                    const to = new Date(2017, JAN, 3);
                     testComponent.moment = [from, to];
                     fixture.detectChanges();
                     flush();
@@ -1499,8 +1565,8 @@ describe('OwlDateTimeComponent', () => {
                     fixture.detectChanges();
                     flush();
 
-                    let newSelectedTo = new Date(2017, JAN, 4);
-                    let containerDebugElement = fixture.debugElement.query(
+                    const newSelectedTo = new Date(2017, JAN, 4);
+                    const containerDebugElement = fixture.debugElement.query(
                         By.directive(OwlDateTimeContainerComponent)
                     );
                     containerDebugElement.componentInstance.dateSelected(
@@ -1530,8 +1596,8 @@ describe('OwlDateTimeComponent', () => {
                 }));
 
                 it('should update toValue and set fromValue to null when date is selected before fromValue', fakeAsync(() => {
-                    let from = new Date(2017, JAN, 2);
-                    let to = new Date(2017, JAN, 3);
+                    const from = new Date(2017, JAN, 2);
+                    const to = new Date(2017, JAN, 3);
                     testComponent.moment = [from, to];
                     fixture.detectChanges();
                     flush();
@@ -1541,8 +1607,8 @@ describe('OwlDateTimeComponent', () => {
                     fixture.detectChanges();
                     flush();
 
-                    let newSelectedTo = new Date(2017, JAN, 1);
-                    let containerDebugElement = fixture.debugElement.query(
+                    const newSelectedTo = new Date(2017, JAN, 1);
+                    const containerDebugElement = fixture.debugElement.query(
                         By.directive(OwlDateTimeContainerComponent)
                     );
                     containerDebugElement.componentInstance.dateSelected(
@@ -1574,11 +1640,11 @@ describe('OwlDateTimeComponent', () => {
         });
 
         describe('DateTimePicker with FormControl', () => {
-            let fixture: ComponentFixture<DateTimePickerWithFormControl>;
-            let testComponent: DateTimePickerWithFormControl;
+            let fixture: ComponentFixture<DateTimePickerWithFormControlComponent>;
+            let testComponent: DateTimePickerWithFormControlComponent;
 
             beforeEach(fakeAsync(() => {
-                fixture = createComponent(DateTimePickerWithFormControl, [
+                fixture = createComponent(DateTimePickerWithFormControlComponent, [
                     OwlNativeDateTimeModule
                 ]);
                 fixture.detectChanges();
@@ -1595,7 +1661,7 @@ describe('OwlDateTimeComponent', () => {
                 expect(testComponent.dateTimePickerInput.value).toBeNull();
                 expect(testComponent.dateTimePicker.selected).toBeNull();
 
-                let selected = new Date(2017, JAN, 1);
+                const selected = new Date(2017, JAN, 1);
                 testComponent.formControl.setValue(selected);
                 fixture.detectChanges();
 
@@ -1609,7 +1675,7 @@ describe('OwlDateTimeComponent', () => {
                 expect(testComponent.formControl.value).toBeNull();
                 expect(testComponent.dateTimePickerInput.value).toBeNull();
 
-                let selected = new Date(2017, JAN, 1);
+                const selected = new Date(2017, JAN, 1);
                 testComponent.dateTimePicker.select(selected);
                 fixture.detectChanges();
                 flush();
@@ -1625,7 +1691,7 @@ describe('OwlDateTimeComponent', () => {
             }));
 
             it('should disable input when form control disabled', () => {
-                let inputEl = fixture.debugElement.query(By.css('input'))
+                const inputEl = fixture.debugElement.query(By.css('input'))
                     .nativeElement;
 
                 expect(inputEl.disabled).toBe(false);
@@ -1647,11 +1713,11 @@ describe('OwlDateTimeComponent', () => {
         });
 
         describe('DateTimePicker with owlDateTimeTrigger', () => {
-            let fixture: ComponentFixture<DateTimePickerWithTrigger>;
-            let testComponent: DateTimePickerWithTrigger;
+            let fixture: ComponentFixture<DateTimePickerWithTriggerComponent>;
+            let testComponent: DateTimePickerWithTriggerComponent;
 
             beforeEach(fakeAsync(() => {
-                fixture = createComponent(DateTimePickerWithTrigger, [
+                fixture = createComponent(DateTimePickerWithTriggerComponent, [
                     OwlNativeDateTimeModule
                 ]);
                 fixture.detectChanges();
@@ -1670,7 +1736,7 @@ describe('OwlDateTimeComponent', () => {
                     document.querySelector('owl-date-time-container')
                 ).toBeNull();
 
-                let toggle = fixture.debugElement.query(By.css('button'));
+                const toggle = fixture.debugElement.query(By.css('button'));
                 dispatchMouseEvent(toggle.nativeElement, 'click');
                 fixture.detectChanges();
 
@@ -1722,15 +1788,15 @@ describe('OwlDateTimeComponent', () => {
 
         describe('DateTimePicker with min and max validation', () => {
             let fixture: ComponentFixture<
-                DateTimePickerWithMinAndMaxValidation
+                DateTimePickerWithMinAndMaxValidationComponent
             >;
-            let testComponent: DateTimePickerWithMinAndMaxValidation;
+            let testComponent: DateTimePickerWithMinAndMaxValidationComponent;
             let minMoment;
             let maxMoment;
 
             beforeEach(fakeAsync(() => {
                 fixture = createComponent(
-                    DateTimePickerWithMinAndMaxValidation,
+                    DateTimePickerWithMinAndMaxValidationComponent,
                     [OwlNativeDateTimeModule]
                 );
                 fixture.detectChanges();
@@ -1830,18 +1896,18 @@ describe('OwlDateTimeComponent', () => {
                 fixture.detectChanges();
                 flush();
                 fixture.detectChanges();
-                let calendarDebugElement = fixture.debugElement.query(
+                const calendarDebugElement = fixture.debugElement.query(
                     By.directive(OwlDateTimeContainerComponent)
                 );
-                let calendarElement = calendarDebugElement.nativeElement;
+                const calendarElement = calendarDebugElement.nativeElement;
 
-                let decreaseHourBtn = calendarElement.querySelector(
+                const decreaseHourBtn = calendarElement.querySelector(
                     '[aria-label="Minus a hour"]'
                 );
-                let decreaseMinuteBtn = calendarElement.querySelector(
+                const decreaseMinuteBtn = calendarElement.querySelector(
                     '[aria-label="Minus a minute"]'
                 );
-                let decreaseSecondBtn = calendarElement.querySelector(
+                const decreaseSecondBtn = calendarElement.querySelector(
                     '[aria-label="Minus a second"]'
                 );
                 expect(decreaseHourBtn.hasAttribute('disabled')).toBe(true);
@@ -1859,18 +1925,18 @@ describe('OwlDateTimeComponent', () => {
                 fixture.detectChanges();
                 flush();
                 fixture.detectChanges();
-                let calendarDebugElement = fixture.debugElement.query(
+                const calendarDebugElement = fixture.debugElement.query(
                     By.directive(OwlDateTimeContainerComponent)
                 );
-                let calendarElement = calendarDebugElement.nativeElement;
+                const calendarElement = calendarDebugElement.nativeElement;
 
-                let increaseHourBtn = calendarElement.querySelector(
+                const increaseHourBtn = calendarElement.querySelector(
                     '[aria-label="Add a hour"]'
                 );
-                let increaseMinuteBtn = calendarElement.querySelector(
+                const increaseMinuteBtn = calendarElement.querySelector(
                     '[aria-label="Add a minute"]'
                 );
-                let increaseSecondBtn = calendarElement.querySelector(
+                const increaseSecondBtn = calendarElement.querySelector(
                     '[aria-label="Add a second"]'
                 );
                 expect(increaseHourBtn.hasAttribute('disabled')).toBe(true);
@@ -1880,11 +1946,11 @@ describe('OwlDateTimeComponent', () => {
         });
 
         describe('DateTimePicker with filter validation', () => {
-            let fixture: ComponentFixture<DateTimePickerWithFilterValidation>;
-            let testComponent: DateTimePickerWithFilterValidation;
+            let fixture: ComponentFixture<DateTimePickerWithFilterValidationComponent>;
+            let testComponent: DateTimePickerWithFilterValidationComponent;
 
             beforeEach(fakeAsync(() => {
-                fixture = createComponent(DateTimePickerWithFilterValidation, [
+                fixture = createComponent(DateTimePickerWithFilterValidationComponent, [
                     OwlNativeDateTimeModule
                 ]);
                 fixture.detectChanges();
@@ -1934,10 +2000,10 @@ describe('OwlDateTimeComponent', () => {
                     document.querySelector('owl-date-time-container')
                 ).not.toBeNull();
 
-                let cellOne = document.querySelector(
+                const cellOne = document.querySelector(
                     '[aria-label="January 1, 2017"]'
                 );
-                let cellTwo = document.querySelector(
+                const cellTwo = document.querySelector(
                     '[aria-label="January 2, 2017"]'
                 );
 
@@ -1952,14 +2018,14 @@ describe('OwlDateTimeComponent', () => {
 
         describe('DateTimePicker with change and input events', () => {
             let fixture: ComponentFixture<
-                DateTimePickerWithChangeAndInputEvents
+                DateTimePickerWithChangeAndInputEventsComponent
             >;
-            let testComponent: DateTimePickerWithChangeAndInputEvents;
+            let testComponent: DateTimePickerWithChangeAndInputEventsComponent;
             let inputEl: HTMLInputElement;
 
             beforeEach(fakeAsync(() => {
                 fixture = createComponent(
-                    DateTimePickerWithChangeAndInputEvents,
+                    DateTimePickerWithChangeAndInputEventsComponent,
                     [OwlNativeDateTimeModule]
                 );
                 fixture.detectChanges();
@@ -2234,11 +2300,11 @@ describe('OwlDateTimeComponent', () => {
         });
 
         describe('DateTimePicker with ISO strings', () => {
-            let fixture: ComponentFixture<DateTimePickerWithISOStrings>;
-            let testComponent: DateTimePickerWithISOStrings;
+            let fixture: ComponentFixture<DateTimePickerWithISOStringsComponent>;
+            let testComponent: DateTimePickerWithISOStringsComponent;
 
             beforeEach(fakeAsync(() => {
-                fixture = createComponent(DateTimePickerWithISOStrings, [
+                fixture = createComponent(DateTimePickerWithISOStringsComponent, [
                     OwlNativeDateTimeModule
                 ]);
                 fixture.detectChanges();
@@ -2272,11 +2338,11 @@ describe('OwlDateTimeComponent', () => {
         });
 
         describe('DateTimePicker with events', () => {
-            let fixture: ComponentFixture<DateTimePickerWithEvents>;
-            let testComponent: DateTimePickerWithEvents;
+            let fixture: ComponentFixture<DateTimePickerWithEventsComponent>;
+            let testComponent: DateTimePickerWithEventsComponent;
 
             beforeEach(fakeAsync(() => {
-                fixture = createComponent(DateTimePickerWithEvents, [
+                fixture = createComponent(DateTimePickerWithEventsComponent, [
                     OwlNativeDateTimeModule
                 ]);
                 fixture.detectChanges();
@@ -2313,7 +2379,7 @@ describe('OwlDateTimeComponent', () => {
 
     describe('with missing DateTimeAdapter and OWL_DATE_TIME_FORMATS', () => {
         it('should throw when created', () => {
-            expect(() => createComponent(StandardDateTimePicker)).toThrowError(
+            expect(() => createComponent(StandardDateTimePickerComponent)).toThrowError(
                 /OwlDateTimePicker: No provider found for .*/
             );
         });
@@ -2329,7 +2395,7 @@ describe('OwlDateTimeComponent', () => {
                        [pickerMode]="pickerMode" #dt></owl-date-time>
     `
 })
-class StandardDateTimePicker {
+class StandardDateTimePickerComponent {
     date: Date | null = new Date(2020, JAN, 1);
     pickerType = 'both';
     pickerMode = 'popup';
@@ -2344,15 +2410,16 @@ class StandardDateTimePicker {
 @Component({
     template: `
         <input [owlDateTime]="dt" [selectMode]="selectMode" [values]="dates">
-        <owl-date-time [startAt]="startAt"
+        <owl-date-time [startAt]="startAt" [endAt]="endAt"
                        [pickerType]="pickerType" #dt></owl-date-time>
     `
 })
-class RangeDateTimePicker {
-    dates: Date[] | null = [new Date(2020, JAN, 1), new Date(2020, FEB, 1)];
+class RangeDateTimePickerComponent {
+     dates: Date[] | null = [new Date(2020, JAN, 1), new Date(2020, FEB, 1)];
     selectMode = 'range';
     pickerType = 'both';
     startAt = new Date(2020, JAN, 1);
+    endAt = new Date(2020, JAN, 2);
     @ViewChild('dt', { static: true })
     dateTimePicker: OwlDateTimeComponent<Date>;
     @ViewChild(OwlDateTimeInputDirective, { static: true })
@@ -2366,14 +2433,14 @@ class RangeDateTimePicker {
         <owl-date-time #dt></owl-date-time>
     `
 })
-class MultiInputDateTimePicker {}
+class MultiInputDateTimePickerComponent { }
 
 @Component({
     template: `
         <owl-date-time #dt></owl-date-time>
     `
 })
-class NoInputDateTimePicker {
+class NoInputDateTimePickerComponent {
     @ViewChild('dt', { static: true })
     dateTimePicker: OwlDateTimeComponent<Date>;
 }
@@ -2384,7 +2451,7 @@ class NoInputDateTimePicker {
         <owl-date-time #dt [startAt]="startDate"></owl-date-time>
     `
 })
-class DateTimePickerWithStartAt {
+class DateTimePickerWithStartAtComponent {
     date = new Date(2020, JAN, 1);
     startDate = new Date(2010, JAN, 1);
     @ViewChild('dt', { static: true })
@@ -2399,15 +2466,15 @@ class DateTimePickerWithStartAt {
                        (yearSelected)="onYearSelection()"></owl-date-time>
     `
 })
-class DateTimePickerWithStartView {
+class DateTimePickerWithStartViewComponent {
     date = new Date(2020, JAN, 1);
     startView = 'month';
     @ViewChild('dt', { static: true })
     dateTimePicker: OwlDateTimeComponent<Date>;
 
-    onMonthSelection() {}
+    onMonthSelection() { }
 
-    onYearSelection() {}
+    onYearSelection() { }
 }
 
 @Component({
@@ -2416,7 +2483,7 @@ class DateTimePickerWithStartView {
         <owl-date-time #dt></owl-date-time>
     `
 })
-class DateTimePickerWithNgModel {
+class DateTimePickerWithNgModelComponent {
     moment: Date[] | Date | null = null;
     selectMode = 'single';
     @ViewChild('dt', { static: true })
@@ -2433,8 +2500,8 @@ class DateTimePickerWithNgModel {
         <owl-date-time #dt></owl-date-time>
     `
 })
-class DateTimePickerWithFormControl {
-    formControl = new FormControl();
+class DateTimePickerWithFormControlComponent {
+    formControl = new UntypedFormControl();
     @ViewChild('dt', { static: true })
     dateTimePicker: OwlDateTimeComponent<Date>;
     @ViewChild(OwlDateTimeInputDirective, { static: true })
@@ -2450,7 +2517,7 @@ class DateTimePickerWithFormControl {
         <owl-date-time #dt></owl-date-time>
     `
 })
-class DateTimePickerWithTrigger {
+class DateTimePickerWithTriggerComponent {
     @ViewChild('dt', { static: true })
     dateTimePicker: OwlDateTimeComponent<Date>;
     @ViewChild(OwlDateTimeInputDirective, { static: true })
@@ -2465,7 +2532,7 @@ class DateTimePickerWithTrigger {
         <owl-date-time [showSecondsTimer]="true" #dt></owl-date-time>
     `
 })
-class DateTimePickerWithMinAndMaxValidation {
+class DateTimePickerWithMinAndMaxValidationComponent {
     @ViewChild('dt', { static: true })
     dateTimePicker: OwlDateTimeComponent<Date>;
     @ViewChild(OwlDateTimeInputDirective, { static: true })
@@ -2487,7 +2554,7 @@ class DateTimePickerWithMinAndMaxValidation {
         <owl-date-time [showSecondsTimer]="true" #dt></owl-date-time>
     `
 })
-class DateTimePickerWithFilterValidation {
+class DateTimePickerWithFilterValidationComponent {
     @ViewChild('dt', { static: true })
     dateTimePicker: OwlDateTimeComponent<Date>;
     @ViewChild(OwlDateTimeInputDirective, { static: true })
@@ -2495,7 +2562,7 @@ class DateTimePickerWithFilterValidation {
     @ViewChild(OwlDateTimeTriggerDirective, { static: true })
     dateTimePickerToggle: OwlDateTimeTriggerDirective<Date>;
     date: Date;
-    filter = (date: Date) => date.getDate() != 1;
+    filter = (date: Date) => date.getDate() !== 1;
 }
 
 @Component({
@@ -2509,7 +2576,7 @@ class DateTimePickerWithFilterValidation {
         <owl-date-time [showSecondsTimer]="true" #dt></owl-date-time>
     `
 })
-class DateTimePickerWithChangeAndInputEvents {
+class DateTimePickerWithChangeAndInputEventsComponent {
     @ViewChild('dt', { static: true })
     dateTimePicker: OwlDateTimeComponent<Date>;
     @ViewChild(OwlDateTimeInputDirective, { static: true })
@@ -2517,13 +2584,13 @@ class DateTimePickerWithChangeAndInputEvents {
     @ViewChild(OwlDateTimeTriggerDirective, { static: true })
     dateTimePickerToggle: OwlDateTimeTriggerDirective<Date>;
 
-    handleChange() {}
+    handleChange() { }
 
-    handleInput() {}
+    handleInput() { }
 
-    handleDateTimeChange() {}
+    handleDateTimeChange() { }
 
-    handleDateTimeInput() {}
+    handleDateTimeInput() { }
 }
 
 @Component({
@@ -2532,7 +2599,7 @@ class DateTimePickerWithChangeAndInputEvents {
     <owl-date-time #dt [startAt]="startAt"></owl-date-time>
   `
 })
-class DateTimePickerWithISOStrings {
+class DateTimePickerWithISOStringsComponent {
     value = new Date(2017, JUN, 1).toISOString();
     min = new Date(2017, JAN, 1).toISOString();
     max = new Date(2017, DEC, 31).toISOString();
@@ -2549,7 +2616,7 @@ class DateTimePickerWithISOStrings {
     <owl-date-time (afterPickerOpen)="openedSpy()" (afterPickerClosed)="closedSpy()" #dt></owl-date-time>
   `
 })
-class DateTimePickerWithEvents {
+class DateTimePickerWithEventsComponent {
     selected: Date | null = null;
     openedSpy = jasmine.createSpy('opened spy');
     closedSpy = jasmine.createSpy('closed spy');
