@@ -29,8 +29,8 @@ export abstract class DateTimeAdapter<T> {
     protected locale: any;
 
     /** A stream that emits when the locale changes. */
-    protected _localeChanges = new Subject<void>();
-    get localeChanges(): Observable<void> {
+    protected _localeChanges = new Subject<string>();
+    get localeChanges(): Observable<string> {
         return this._localeChanges;
     }
 
@@ -283,9 +283,16 @@ export abstract class DateTimeAdapter<T> {
     /**
      * Sets the locale used for all dates.
      */
-    setLocale(locale: any) {
+    setLocale(locale: string) {
         this.locale = locale;
-        this._localeChanges.next();
+        this._localeChanges.next(locale);
+    }
+
+     /**
+     * Get the locale used for all dates.
+     * */
+    getLocale() {
+        return this.locale;
     }
 
     /**
