@@ -1,8 +1,11 @@
 /**
  * calendar-month-view.component.spec
  */
+import { registerLocaleData } from '@angular/common';
+import localeDutch from '@angular/common/locales/nl';
+import { DateTimeAdapter } from './adapter/date-time-adapter.class';
 import { OwlMonthViewComponent } from './calendar-month-view.component';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { OwlDateTimeIntl } from './date-time-picker-intl.service';
 import { OwlNativeDateTimeModule } from './adapter/native-date-time.module';
 import { OwlDateTimeModule } from './date-time.module';
@@ -35,7 +38,7 @@ const JAN = 0,
     DEC = 11;
 
 describe('OwlMonthViewComponent', () => {
-    beforeEach(async(() => {
+    beforeEach(() => {
         TestBed.configureTestingModule({
             imports: [OwlNativeDateTimeModule, OwlDateTimeModule],
             declarations: [
@@ -44,7 +47,7 @@ describe('OwlMonthViewComponent', () => {
             ],
             providers: [OwlDateTimeIntl]
         }).compileComponents();
-    }));
+    });
 
     describe('standard month view', () => {
         let fixture: ComponentFixture<StandardMonthViewComponent>;
@@ -66,38 +69,38 @@ describe('OwlMonthViewComponent', () => {
         });
 
         it('should have 42 calendar day cells', () => {
-            let cellEls = monthViewElement.querySelectorAll(
+            const cellEls = monthViewElement.querySelectorAll(
                 '.owl-dt-calendar-cell'
-            )!;
+            );
             expect(cellEls.length).toBe(42);
         });
 
         it('should show selected date if in same month', () => {
-            let selectedEl = monthViewElement.querySelector(
+            const selectedEl = monthViewElement.querySelector(
                 '.owl-dt-calendar-cell-selected'
-            )!;
+            );
             expect(selectedEl.innerHTML.trim()).toBe('10');
         });
 
         it('should NOT show selected date if in different month', () => {
             testComponent.selected = new Date(2018, MAR, 10);
             fixture.detectChanges();
-            let selectedEl = monthViewElement.querySelector(
+            const selectedEl = monthViewElement.querySelector(
                 '.owl-dt-calendar-cell-selected'
-            )!;
+            );
             expect(selectedEl).toBeNull();
         });
 
         it('should mark active date', () => {
-            let selectedElContent = monthViewElement.querySelector(
+            const selectedElContent = monthViewElement.querySelector(
                 '.owl-dt-calendar-cell-active .owl-dt-calendar-cell-content'
-            )!;
+            );
             expect(selectedElContent.innerHTML.trim()).toBe('5');
         });
 
         it('should set Sunday as first day of week by default', () => {
             expect(monthViewInstance.firstDayOfWeek).toBe(0);
-            let weekdayCells = monthViewElement.querySelectorAll(
+            const weekdayCells = monthViewElement.querySelectorAll(
                 '.owl-dt-weekday'
             );
             expect(weekdayCells[0].getAttribute('aria-label')).toBe('Sunday');
@@ -106,7 +109,7 @@ describe('OwlMonthViewComponent', () => {
         it('should set Monday as first day of week if firstDayOfWeek set to 1', () => {
             monthViewInstance.firstDayOfWeek = 1;
             fixture.detectChanges();
-            let weekdayCells = monthViewElement.querySelectorAll(
+            const weekdayCells = monthViewElement.querySelectorAll(
                 '.owl-dt-weekday'
             );
             expect(weekdayCells[0].getAttribute('aria-label')).toBe('Monday');
@@ -115,14 +118,14 @@ describe('OwlMonthViewComponent', () => {
         it('should set Saturday as first day of week if firstDayOfWeek set to 6', () => {
             monthViewInstance.firstDayOfWeek = 6;
             fixture.detectChanges();
-            let weekdayCells = monthViewElement.querySelectorAll(
+            const weekdayCells = monthViewElement.querySelectorAll(
                 '.owl-dt-weekday'
             );
             expect(weekdayCells[0].getAttribute('aria-label')).toBe('Saturday');
         });
 
         it('should decrement date on left arrow press', () => {
-            let calendarBodyEl = monthViewElement.querySelector(
+            const calendarBodyEl = monthViewElement.querySelector(
                 '.owl-dt-calendar-body'
             );
             dispatchKeyboardEvent(calendarBodyEl, 'keydown', LEFT_ARROW);
@@ -139,7 +142,7 @@ describe('OwlMonthViewComponent', () => {
         });
 
         it('should increment date on right arrow press', () => {
-            let calendarBodyEl = monthViewElement.querySelector(
+            const calendarBodyEl = monthViewElement.querySelector(
                 '.owl-dt-calendar-body'
             );
             dispatchKeyboardEvent(calendarBodyEl, 'keydown', RIGHT_ARROW);
@@ -154,7 +157,7 @@ describe('OwlMonthViewComponent', () => {
         });
 
         it('should go up a row on up arrow press', () => {
-            let calendarBodyEl = monthViewElement.querySelector(
+            const calendarBodyEl = monthViewElement.querySelector(
                 '.owl-dt-calendar-body'
             );
             dispatchKeyboardEvent(calendarBodyEl, 'keydown', UP_ARROW);
@@ -172,7 +175,7 @@ describe('OwlMonthViewComponent', () => {
         });
 
         it('should go down a row on down arrow press', () => {
-            let calendarBodyEl = monthViewElement.querySelector(
+            const calendarBodyEl = monthViewElement.querySelector(
                 '.owl-dt-calendar-body'
             );
             dispatchKeyboardEvent(calendarBodyEl, 'keydown', DOWN_ARROW);
@@ -190,7 +193,7 @@ describe('OwlMonthViewComponent', () => {
             monthViewInstance.pickerMoment = new Date(2018, JAN, 7);
             fixture.detectChanges();
 
-            let calendarBodyEl = monthViewElement.querySelector(
+            const calendarBodyEl = monthViewElement.querySelector(
                 '.owl-dt-calendar-body'
             );
             dispatchKeyboardEvent(calendarBodyEl, 'keydown', HOME);
@@ -208,7 +211,7 @@ describe('OwlMonthViewComponent', () => {
             monthViewInstance.pickerMoment = new Date(2018, JAN, 7);
             fixture.detectChanges();
 
-            let calendarBodyEl = monthViewElement.querySelector(
+            const calendarBodyEl = monthViewElement.querySelector(
                 '.owl-dt-calendar-body'
             );
             dispatchKeyboardEvent(calendarBodyEl, 'keydown', END);
@@ -223,7 +226,7 @@ describe('OwlMonthViewComponent', () => {
         });
 
         it('should go back one month on page up press', () => {
-            let calendarBodyEl = monthViewElement.querySelector(
+            const calendarBodyEl = monthViewElement.querySelector(
                 '.owl-dt-calendar-body'
             );
             dispatchKeyboardEvent(calendarBodyEl, 'keydown', PAGE_UP);
@@ -238,7 +241,7 @@ describe('OwlMonthViewComponent', () => {
         });
 
         it('should go forward one month on page down press', () => {
-            let calendarBodyEl = monthViewElement.querySelector(
+            const calendarBodyEl = monthViewElement.querySelector(
                 '.owl-dt-calendar-body'
             );
             dispatchKeyboardEvent(calendarBodyEl, 'keydown', PAGE_DOWN);
@@ -253,7 +256,7 @@ describe('OwlMonthViewComponent', () => {
         });
 
         it('should select active date on enter', () => {
-            let calendarBodyEl = monthViewElement.querySelector(
+            const calendarBodyEl = monthViewElement.querySelector(
                 '.owl-dt-calendar-body'
             );
             dispatchKeyboardEvent(calendarBodyEl, 'keydown', LEFT_ARROW);
@@ -276,17 +279,17 @@ describe('OwlMonthViewComponent', () => {
             fixture = TestBed.createComponent(MonthViewWithDateFilterComponent);
             fixture.detectChanges();
 
-            let monthViewDebugElement = fixture.debugElement.query(
+            const monthViewDebugElement = fixture.debugElement.query(
                 By.directive(OwlMonthViewComponent)
             );
             monthViewNativeElement = monthViewDebugElement.nativeElement;
         });
 
         it('should disable filtered dates', () => {
-            let cellOne = monthViewNativeElement.querySelector(
+            const cellOne = monthViewNativeElement.querySelector(
                 '[aria-label="January 1, 2018"]'
             );
-            let cellTwo = monthViewNativeElement.querySelector(
+            const cellTwo = monthViewNativeElement.querySelector(
                 '[aria-label="January 2, 2018"]'
             );
             expect(cellOne.classList).toContain(
@@ -295,6 +298,68 @@ describe('OwlMonthViewComponent', () => {
             expect(cellTwo.classList).not.toContain(
                 'owl-dt-calendar-cell-disabled'
             );
+        });
+    });
+
+
+    describe('standard month view (locale tests)', () => {
+        let fixture: ComponentFixture<StandardMonthViewComponent>;
+        let adapter: DateTimeAdapter<unknown>;
+        let monthViewDebugElement: DebugElement;
+        let monthViewElement: HTMLElement;
+        let monthViewInstance: OwlMonthViewComponent<Date>;
+
+        beforeAll(() => {
+            registerLocaleData(localeDutch);
+        });
+
+        beforeEach(() => {
+            fixture = TestBed.createComponent(StandardMonthViewComponent);
+
+            adapter = TestBed.inject(DateTimeAdapter);
+            monthViewDebugElement = fixture.debugElement.query(
+                By.directive(OwlMonthViewComponent)
+            );
+            monthViewElement = monthViewDebugElement.nativeElement;
+            monthViewInstance = monthViewDebugElement.componentInstance;
+        });
+
+        it('should derive the first day of the week based on the active locale', () => {
+            adapter.setLocale('nl-NL');
+
+            fixture.detectChanges();
+            const weekdayCells = monthViewElement.querySelectorAll(
+                '.owl-dt-weekday'
+            );
+            expect(weekdayCells[0].getAttribute('aria-label')).toBe('maandag');
+        });
+
+        it('should fallback to Sunday as first day of the week when when locale data is missing', () => {
+            adapter.setLocale('unknown');
+
+            fixture.detectChanges();
+            const weekdayCells = monthViewElement.querySelectorAll(
+                '.owl-dt-weekday'
+            );
+            expect(weekdayCells[0].getAttribute('aria-label')).toBe('Sunday');
+        });
+
+        it('should update the default day of the week when locale changes', () => {
+            adapter.setLocale('nl-NL');
+
+            fixture.detectChanges();
+            const weekdayCellsNl = monthViewElement.querySelectorAll(
+                '.owl-dt-weekday'
+            );
+            expect(weekdayCellsNl[0].getAttribute('aria-label')).toBe('maandag');
+
+            adapter.setLocale('en-US');
+
+            fixture.detectChanges();
+            const weekdayCellsUs = monthViewElement.querySelectorAll(
+                '.owl-dt-weekday'
+            );
+            expect(weekdayCellsUs[0].getAttribute('aria-label')).toBe('Sunday');
         });
     });
 });
@@ -321,6 +386,7 @@ class StandardMonthViewComponent {
 class MonthViewWithDateFilterComponent {
     pickerMoment = new Date(2018, JAN, 1);
     dateFilter(date: Date) {
-        return date.getDate() % 2 == 0;
+      // tslint:disable-next-line:triple-equals
+        return date.getDate() % 2 === 0;
     }
 }
