@@ -38,7 +38,7 @@ import {
     RIGHT_ARROW,
     UP_ARROW
 } from '@angular/cdk/keycodes';
-import { coerceNumberProperty } from '@angular/cdk/coercion';
+import { getLocaleFirstDayOfWeek } from '@angular/common';
 
 const DAYS_PER_WEEK = 7;
 const WEEKS_PER_VIEW = 6;
@@ -60,22 +60,25 @@ export class OwlMonthViewComponent<T>
      * Whether to hide dates in other months at the start or end of the current month.
      * */
     @Input()
-    hideOtherMonths: boolean = false;
+    hideOtherMonths = false;
+
+    private isDefaultFirstDayOfWeek = true;
 
     /**
      * Define the first day of a week
-     * Sunday: 0 ~ Saturday: 6
+     * Sunday: 0 - Saturday: 6
      * */
-    private _firstDayOfWeek: number = 0;
+    private _firstDayOfWeek: number;
+
     @Input()
     get firstDayOfWeek(): number {
         return this._firstDayOfWeek;
     }
 
     set firstDayOfWeek(val: number) {
-        val = coerceNumberProperty(val);
         if (val >= 0 && val <= 6 && val !== this._firstDayOfWeek) {
             this._firstDayOfWeek = val;
+            this.isDefaultFirstDayOfWeek = false;
 
             if (this.initiated) {
                 this.generateWeekDays();
@@ -296,13 +299,17 @@ export class OwlMonthViewComponent<T>
     ) {}
 
     public ngOnInit() {
+        this.updateFirstDayOfWeek(this.dateTimeAdapter.getLocale());
         this.generateWeekDays();
 
-        this.localeSub = this.dateTimeAdapter.localeChanges.subscribe(() => {
-            this.generateWeekDays();
-            this.generateCalendar();
-            this.cdRef.markForCheck();
-        });
+        this.localeSub = this.dateTimeAdapter.localeChanges.subscribe(
+            locale => {
+                this.updateFirstDayOfWeek(locale);
+                this.generateWeekDays();
+                this.generateCalendar();
+                this.cdRef.markForCheck();
+            }
+        );
     }
 
     public ngAfterContentInit(): void {
@@ -522,6 +529,16 @@ export class OwlMonthViewComponent<T>
         }
 
         this.setSelectedDates();
+    }
+
+    private updateFirstDayOfWeek(locale: string): void {
+        if (this.isDefaultFirstDayOfWeek) {
+            try {
+                this._firstDayOfWeek = getLocaleFirstDayOfWeek(locale);
+            } catch {
+                this._firstDayOfWeek = 0;
+            }
+        }
     }
 
     /**
