@@ -7,6 +7,7 @@ import {
     ChangeDetectionStrategy, ChangeDetectorRef,
     Component,
     EventEmitter,
+    Inject,
     Input,
     OnInit,
     Optional,
@@ -28,15 +29,13 @@ import {
     UP_ARROW
 } from '@angular/cdk/keycodes';
 import { OwlDateTimeIntl } from './date-time-picker-intl.service';
-
-export const YEARS_PER_ROW = 3;
-export const YEAR_ROWS = 7;
+import { OptionsTokens, Options } from './options-provider';
 
 @Component({
     selector: 'owl-date-time-multi-year-view',
     templateUrl: './calendar-multi-year-view.component.html',
     styleUrls: ['./calendar-multi-year-view.component.scss'],
-    host:{
+    host: {
         '[class.owl-dt-calendar-view]': 'owlDTCalendarView',
         '[class.owl-dt-calendar-multi-year-view]': 'owlDTCalendarMultiYearView'
     },
@@ -185,13 +184,13 @@ export class OwlMultiYearViewComponent<T> implements OnInit, AfterContentInit {
 
     get activeCell(): number {
         if (this._pickerMoment) {
-            return this.dateTimeAdapter.getYear(this._pickerMoment) % (YEARS_PER_ROW * YEAR_ROWS);
+            return this.dateTimeAdapter.getYear(this._pickerMoment) % (this.options.yearsPerRow * this.options.yearRows);
         }
     }
 
     get tableHeader(): string {
         if (this._years && this._years.length > 0) {
-            return `${this._years[0][0].displayValue} ~ ${this._years[YEAR_ROWS - 1][YEARS_PER_ROW - 1].displayValue}`
+            return `${this._years[0][0].displayValue} - ${this._years[this.options.yearRows - 1][this.options.yearsPerRow - 1].displayValue}`;
         }
     }
 
@@ -232,7 +231,8 @@ export class OwlMultiYearViewComponent<T> implements OnInit, AfterContentInit {
 
     constructor( private cdRef: ChangeDetectorRef,
                  private pickerIntl: OwlDateTimeIntl,
-                 @Optional() private dateTimeAdapter: DateTimeAdapter<T> ) {
+                 @Optional() private dateTimeAdapter: DateTimeAdapter<T>,
+                 @Inject(OptionsTokens.multiYear) private options: any) {
     }
 
     public ngOnInit() {
@@ -275,7 +275,7 @@ export class OwlMultiYearViewComponent<T> implements OnInit, AfterContentInit {
      * Generate the previous year list
      * */
     public prevYearList( event: any ): void {
-        this._pickerMoment = this.dateTimeAdapter.addCalendarYears(this.pickerMoment, -1 * YEAR_ROWS * YEARS_PER_ROW);
+        this._pickerMoment = this.dateTimeAdapter.addCalendarYears(this.pickerMoment, -1 * this.options.yearsPerRow * this.options.yearRows);
         this.generateYearList();
         event.preventDefault();
     }
@@ -284,7 +284,7 @@ export class OwlMultiYearViewComponent<T> implements OnInit, AfterContentInit {
      * Generate the next year list
      * */
     public nextYearList( event: any ): void {
-        this._pickerMoment = this.dateTimeAdapter.addCalendarYears(this.pickerMoment, YEAR_ROWS * YEARS_PER_ROW);
+        this._pickerMoment = this.dateTimeAdapter.addCalendarYears(this.pickerMoment, this.options.yearsPerRow * this.options.yearRows);
         this.generateYearList();
         event.preventDefault();
     }
@@ -293,13 +293,13 @@ export class OwlMultiYearViewComponent<T> implements OnInit, AfterContentInit {
         this._years = [];
 
         const pickerMomentYear = this.dateTimeAdapter.getYear(this._pickerMoment);
-        const offset = pickerMomentYear % (YEARS_PER_ROW * YEAR_ROWS);
+        const offset = pickerMomentYear % (this.options.yearsPerRow * this.options.yearRows);
 
-        for (let i = 0; i < YEAR_ROWS; i++) {
+        for (let i = 0; i < this.options.yearRows; i++) {
             const row = [];
 
-            for (let j = 0; j < YEARS_PER_ROW; j++) {
-                const year = pickerMomentYear - offset + (j + i * YEARS_PER_ROW);
+            for (let j = 0; j < this.options.yearsPerRow; j++) {
+                const year = pickerMomentYear - offset + (j + i * this.options.yearsPerRow);
                 const yearCell = this.createYearCell(year);
                 row.push(yearCell);
             }
@@ -341,39 +341,39 @@ export class OwlMultiYearViewComponent<T> implements OnInit, AfterContentInit {
 
             // minus 3 years
             case UP_ARROW:
-                moment = this.dateTimeAdapter.addCalendarYears(this._pickerMoment, -1 * YEARS_PER_ROW);
+                moment = this.dateTimeAdapter.addCalendarYears(this._pickerMoment, -1 * this.options.yearsPerRow);
                 this.pickerMomentChange.emit(moment);
                 break;
 
             // add 3 years
             case DOWN_ARROW:
-                moment = this.dateTimeAdapter.addCalendarYears(this._pickerMoment, YEARS_PER_ROW);
+                moment = this.dateTimeAdapter.addCalendarYears(this._pickerMoment, this.options.yearsPerRow);
                 this.pickerMomentChange.emit(moment);
                 break;
 
             // go to the first year of the year page
             case HOME:
                 moment = this.dateTimeAdapter.addCalendarYears(this._pickerMoment,
-                    -this.dateTimeAdapter.getYear(this._pickerMoment) % (YEARS_PER_ROW * YEAR_ROWS));
+                    -this.dateTimeAdapter.getYear(this._pickerMoment) % (this.options.yearsPerRow * this.options.yearRows));
                 this.pickerMomentChange.emit(moment);
                 break;
 
             // go to the last year of the year page
             case END:
                 moment = this.dateTimeAdapter.addCalendarYears(this._pickerMoment,
-                    (YEARS_PER_ROW * YEAR_ROWS) - this.dateTimeAdapter.getYear(this._pickerMoment) % (YEARS_PER_ROW * YEAR_ROWS) - 1);
+                    (this.options.yearsPerRow * this.options.yearRows) - this.dateTimeAdapter.getYear(this._pickerMoment) % (this.options.yearsPerRow * this.options.yearRows) - 1);
                 this.pickerMomentChange.emit(moment);
                 break;
 
             // minus 1 year page (or 10 year pages)
             case PAGE_UP:
-                moment = this.dateTimeAdapter.addCalendarYears(this.pickerMoment, event.altKey ? -10 * (YEARS_PER_ROW * YEAR_ROWS) : -1 * (YEARS_PER_ROW * YEAR_ROWS));
+                moment = this.dateTimeAdapter.addCalendarYears(this.pickerMoment, event.altKey ? -10 * (this.options.yearsPerRow * this.options.yearRows) : -1 * (this.options.yearsPerRow * this.options.yearRows));
                 this.pickerMomentChange.emit(moment);
                 break;
 
             // add 1 year page (or 10 year pages)
             case PAGE_DOWN:
-                moment = this.dateTimeAdapter.addCalendarYears(this.pickerMoment, event.altKey ? 10 * (YEARS_PER_ROW * YEAR_ROWS) : (YEARS_PER_ROW * YEAR_ROWS));
+                moment = this.dateTimeAdapter.addCalendarYears(this.pickerMoment, event.altKey ? 10 * (this.options.yearsPerRow * this.options.yearRows) : (this.options.yearsPerRow * this.options.yearRows));
                 this.pickerMomentChange.emit(moment);
                 break;
 
@@ -415,7 +415,7 @@ export class OwlMultiYearViewComponent<T> implements OnInit, AfterContentInit {
                 } else {
                     return null;
                 }
-            })
+            });
         }
     }
 
@@ -436,7 +436,7 @@ export class OwlMultiYearViewComponent<T> implements OnInit, AfterContentInit {
         const firstOfYear = this.dateTimeAdapter.createDate(year, 0, 1);
 
         // If any date in the year is enabled count the year as enabled.
-        for (let date = firstOfYear; this.dateTimeAdapter.getYear(date) == year;
+        for (let date = firstOfYear; this.dateTimeAdapter.getYear(date) === year;
              date = this.dateTimeAdapter.addCalendarDays(date, 1)) {
             if (this.dateFilter(date)) {
                 return true;
@@ -447,8 +447,8 @@ export class OwlMultiYearViewComponent<T> implements OnInit, AfterContentInit {
     }
 
     private isSameYearList( date1: T, date2: T ): boolean {
-        return Math.floor(this.dateTimeAdapter.getYear(date1) / (YEARS_PER_ROW * YEAR_ROWS)) ===
-            Math.floor(this.dateTimeAdapter.getYear(date2) / (YEARS_PER_ROW * YEAR_ROWS));
+        return Math.floor(this.dateTimeAdapter.getYear(date1) / (this.options.yearsPerRow * this.options.yearRows)) ===
+            Math.floor(this.dateTimeAdapter.getYear(date2) / (this.options.yearsPerRow * this.options.yearRows));
     }
 
     /**
