@@ -126,6 +126,32 @@ export class OwlDateTimeInlineComponent<T> extends OwlDateTime<T>
         );
     }
 
+    /** The date to open for range calendar. */
+    private _endAt: T | null;
+    @Input()
+    get endAt(): T | null {
+        if (this._endAt) {
+            return this._endAt;
+        }
+
+        if (this.selectMode === 'single') {
+            return this.value || null;
+        } else if (
+            this.selectMode === 'range' ||
+            this.selectMode === 'rangeFrom'
+        ) {
+            return this.values[1] || null;
+        } else {
+            return null;
+        }
+    }
+
+    set endAt(date: T | null) {
+        this._endAt = this.getValidDate(
+            this.dateTimeAdapter.deserialize(date)
+        );
+    }
+
     private _dateTimeFilter: (date: T | null) => boolean;
     @Input('owlDateTimeFilter')
     get dateTimeFilter() {
@@ -210,6 +236,12 @@ export class OwlDateTimeInlineComponent<T> extends OwlDateTime<T>
     @Output()
     monthSelected = new EventEmitter<T>();
 
+    /**
+     * Emits selected date
+     * */
+    @Output()
+    dateSelected = new EventEmitter<T>();
+
     private _selected: T | null;
     get selected() {
         return this._selected;
@@ -254,8 +286,8 @@ export class OwlDateTimeInlineComponent<T> extends OwlDateTime<T>
         return true;
     }
 
-    private onModelChange: Function = () => {};
-    private onModelTouched: Function = () => {};
+    private onModelChange: Function = () => { };
+    private onModelTouched: Function = () => { };
 
     constructor(
         protected changeDetector: ChangeDetectorRef,
@@ -321,5 +353,12 @@ export class OwlDateTimeInlineComponent<T> extends OwlDateTime<T>
      * */
     public selectMonth(normalizedMonth: T): void {
         this.monthSelected.emit(normalizedMonth);
+    }
+    
+    /**
+     * Emits the selected date
+     * */
+     public selectDate(normalizedDate: T): void {
+        this.dateSelected.emit(normalizedDate);
     }
 }
