@@ -75,6 +75,22 @@ export class OwlDateTimeInputDirective<T>
         OnDestroy,
         ControlValueAccessor,
         Validator {
+    static ngAcceptInputType_disabled: boolean|'';
+
+     /**
+     * Required flag to be used for range of [null, null]
+     * */
+    private _required: boolean;
+    @Input()
+    get required() {
+        return this._required;
+    }
+
+    set required(value: any) {
+        this._required = value === '' || value;
+        this.validatorOnChange();
+    }
+
     /**
      * The date time picker that this input is associated with.
      * */
@@ -172,7 +188,7 @@ export class OwlDateTimeInputDirective<T>
      * The character to separate the 'from' and 'to' in input value
      */
     @Input()
-    rangeSeparator = '~';
+    rangeSeparator = '-';
 
     private _value: T | null;
     @Input()
@@ -269,7 +285,7 @@ export class OwlDateTimeInputDirective<T>
         return this.lastValueValid
             ? null
             : { owlDateTimeParse: { text: this.elmRef.nativeElement.value } };
-    };
+    }
 
     /** The form control validator for the min date. */
     private minValidator: ValidatorFn = (
@@ -303,7 +319,7 @@ export class OwlDateTimeInputDirective<T>
                       }
                   };
         }
-    };
+    }
 
     /** The form control validator for the max date. */
     private maxValidator: ValidatorFn = (
@@ -337,7 +353,7 @@ export class OwlDateTimeInputDirective<T>
                       }
                   };
         }
-    };
+    }
 
     /** The form control validator for the date filter. */
     private filterValidator: ValidatorFn = (
@@ -351,7 +367,7 @@ export class OwlDateTimeInputDirective<T>
             this._dateTimeFilter(controlValue)
             ? null
             : { owlDateTimeFilter: true };
-    };
+    }
 
     /**
      * The form control validator for the range.
@@ -376,7 +392,31 @@ export class OwlDateTimeInputDirective<T>
             this.dateTimeAdapter.compare(controlValueFrom, controlValueTo) <= 0
             ? null
             : { owlDateTimeRange: true };
-    };
+    }
+
+    /**
+     * The form control validator for the range when required.
+     * Check whether the 'before' and 'to' values are present
+     * */
+    private requiredRangeValidator: ValidatorFn = (
+        control: AbstractControl
+    ): ValidationErrors | null => {
+        if (this.isInSingleMode || !control.value || !this.required) {
+            return null;
+        }
+
+        const controlValueFrom = this.getValidDate(
+            this.dateTimeAdapter.deserialize(control.value[0])
+        );
+        const controlValueTo = this.getValidDate(
+            this.dateTimeAdapter.deserialize(control.value[1])
+        );
+
+        return !controlValueFrom ||
+            !controlValueTo
+            ? { owlRequiredDateTimeRange: [controlValueFrom, controlValueTo] }
+            : null;
+    }
 
     /** The combined form control validator for this input. */
     private validator: ValidatorFn | null = Validators.compose([
@@ -384,7 +424,8 @@ export class OwlDateTimeInputDirective<T>
         this.minValidator,
         this.maxValidator,
         this.filterValidator,
-        this.rangeValidator
+        this.rangeValidator,
+        this.requiredRangeValidator
     ]);
 
     /** Emits when the value changes (either due to user input or programmatic change). */
@@ -413,10 +454,13 @@ export class OwlDateTimeInputDirective<T>
         return this.disabled;
     }
 
-    constructor( private elmRef: ElementRef,
+    constructor(
+        private elmRef: ElementRef,
         private renderer: Renderer2,
-        @Optional() private dateTimeAdapter: DateTimeAdapter<T>,
-        @Optional() @Inject(OWL_DATE_TIME_FORMATS) private dateTimeFormats: OwlDateTimeFormats ) {
+        @Optional()
+        private dateTimeAdapter: DateTimeAdapter<T>,
+        @Optional() @Inject(OWL_DATE_TIME_FORMATS)
+        private dateTimeFormats: OwlDateTimeFormats ) {
         if (!this.dateTimeAdapter) {
             throw Error(
                 `OwlDateTimePicker: No provider found for DateTimePicker. You must import one of the following ` +
@@ -509,19 +553,19 @@ export class OwlDateTimeInputDirective<T>
     /**
      * Open the picker when user hold alt + DOWN_ARROW
      * */
-    public handleKeydownOnHost( event: KeyboardEvent ): void {
+    public handleKeydownOnHost(event: KeyboardEvent ): void {
         if (event.altKey && event.keyCode === DOWN_ARROW) {
             this.dtPicker.open();
             event.preventDefault();
         }
     }
 
-    public handleBlurOnHost( event: Event ): void {
+    public handleBlurOnHost(event: Event ): void {
         this.onModelTouched();
     }
 
-    public handleInputOnHost( event: any ): void {
-        let value = event.target.value;
+    public handleInputOnHost(event: any ): void {
+        const value = event.target.value;
         if (this._selectMode === 'single') {
             this.changeInputInSingleMode(value);
         } else if (this._selectMode === 'range') {
@@ -531,7 +575,7 @@ export class OwlDateTimeInputDirective<T>
         }
     }
 
-    public handleChangeOnHost( event: any ): void {
+    public handleChangeOnHost(event: any ): void {
 
         let v;
         if (this.isInSingleMode) {
@@ -699,7 +743,7 @@ export class OwlDateTimeInputDirective<T>
      * Handle input change in rangeFrom or rangeTo mode
      */
     private changeInputInRangeFromToMode(inputValue: string): void {
-        let originalValue =
+        const originalValue =
             this._selectMode === 'rangeFrom'
                 ? this._values[0]
                 : this._values[1];
@@ -801,6 +845,6 @@ export class OwlDateTimeInputDirective<T>
             return this.dateTimeAdapter.compare(first, second) === 0;
         }
 
-        return first == second;
+        return first === second;
     }
 }
