@@ -45,16 +45,16 @@ export class OwlDateTimeIntl {
     nextYearLabel = 'Next year';
 
     /** A label for the previous multi-year button (used by screen readers). */
-    prevMultiYearLabel: string = 'Previous 21 years';
+    prevMultiYearLabel = 'Previous 21 years';
 
     /** A label for the next multi-year button (used by screen readers). */
-    nextMultiYearLabel: string = 'Next 21 years';
+    nextMultiYearLabel = 'Next 21 years';
 
     /** A label for the 'switch to month view' button (used by screen readers). */
     switchToMonthViewLabel = 'Change to month view';
 
     /** A label for the 'switch to year view' button (used by screen readers). */
-    switchToMultiYearViewLabel: string = 'Choose month and year';
+    switchToMultiYearViewLabel = 'Choose month and year';
 
     /** A label for the cancel button */
     cancelBtnLabel = 'Cancel';
