@@ -118,6 +118,34 @@ export class OwlDateTimeComponent<T> extends OwlDateTime<T>
         );
     }
 
+    /** The end date to set for range calendar. */
+    private _endAt: T | null;
+    @Input()
+    get endAt(): T | null {
+        if (this._endAt) {
+            return this._endAt;
+        }
+
+        if (this._dtInput) {
+            if (this._dtInput.selectMode === 'single') {
+                return this._dtInput.value || null;
+            } else if (
+                this._dtInput.selectMode === 'range' ||
+                this._dtInput.selectMode === 'rangeFrom'
+            ) {
+                return this._dtInput.values[1] || null;
+            }
+        } else {
+            return null;
+        }
+    }
+
+    set endAt(date: T | null) {
+        this._endAt = this.getValidDate(
+            this.dateTimeAdapter.deserialize(date)
+        );
+    }
+
     /**
      * Set the type of the dateTime picker
      *      'both' -- show both calendar and timer
@@ -174,7 +202,7 @@ export class OwlDateTimeComponent<T> extends OwlDateTime<T>
     }
 
     /** Whether the calendar is open. */
-    private _opened: boolean = false;
+    private _opened = false;
     @Input()
     get opened(): boolean {
         return this._opened;
@@ -198,6 +226,12 @@ export class OwlDateTimeComponent<T> extends OwlDateTime<T>
     afterPickerClosed = new EventEmitter<any>();
 
     /**
+     * Callback before the picker is open
+     * */
+    @Output()
+    beforePickerOpen = new EventEmitter<any>();
+
+    /**
      * Callback when the picker is open
      * */
     @Output()
@@ -216,6 +250,12 @@ export class OwlDateTimeComponent<T> extends OwlDateTime<T>
      * */
     @Output()
     monthSelected = new EventEmitter<T>();
+
+    /**
+     * Emits selected date
+     * */
+    @Output()
+    dateSelected = new EventEmitter<T>();
 
     /**
      * Emit when the selected value has been confirmed
@@ -237,6 +277,7 @@ export class OwlDateTimeComponent<T> extends OwlDateTime<T>
     private hidePickerStreamSub = Subscription.EMPTY;
     private confirmSelectedStreamSub = Subscription.EMPTY;
     private pickerOpenedStreamSub = Subscription.EMPTY;
+    private pickerBeforeOpenedStreamSub = Subscription.EMPTY;
 
     /** The element that was focused before the date time picker was opened. */
     private focusedElementBeforeOpen: HTMLElement | null = null;
@@ -292,10 +333,10 @@ export class OwlDateTimeComponent<T> extends OwlDateTime<T>
         return this._dtInput.isInRangeMode;
     }
 
-    private defaultScrollStrategy: () => ScrollStrategy;
+    private readonly defaultScrollStrategy: () => ScrollStrategy;
 
     constructor(
-        private overlay: Overlay,
+        public overlay: Overlay,
         private viewContainerRef: ViewContainerRef,
         private dialogService: OwlDialogService,
         private ngZone: NgZone,
@@ -446,6 +487,13 @@ export class OwlDateTimeComponent<T> extends OwlDateTime<T>
     }
 
     /**
+     * Emits the selected date
+     * */
+     public selectDate(normalizedDate: T): void {
+        this.dateSelected.emit(normalizedDate);
+    }
+
+    /**
      * Hide the picker
      */
     public close(): void {
@@ -474,6 +522,11 @@ export class OwlDateTimeComponent<T> extends OwlDateTime<T>
             this.confirmSelectedStreamSub = null;
         }
 
+        if (this.pickerBeforeOpenedStreamSub) {
+            this.pickerBeforeOpenedStreamSub.unsubscribe();
+            this.pickerBeforeOpenedStreamSub = null;
+        }
+
         if (this.pickerOpenedStreamSub) {
             this.pickerOpenedStreamSub.unsubscribe();
             this.pickerOpenedStreamSub = null;
@@ -487,7 +540,8 @@ export class OwlDateTimeComponent<T> extends OwlDateTime<T>
         const completeClose = () => {
             if (this._opened) {
                 this._opened = false;
-                this.afterPickerClosed.emit(null);
+                const selected = this.selected || this.selecteds;
+                this.afterPickerClosed.emit(selected);
                 this.focusedElementBeforeOpen = null;
             }
         };
@@ -544,6 +598,9 @@ export class OwlDateTimeComponent<T> extends OwlDateTime<T>
         );
         this.pickerContainer = this.dialogRef.componentInstance;
 
+        this.dialogRef.beforeOpen().subscribe(() => {
+            this.beforePickerOpen.emit(null);
+        });
         this.dialogRef.afterOpen().subscribe(() => {
             this.afterPickerOpen.emit(null);
             this._opened = true;
@@ -577,6 +634,12 @@ export class OwlDateTimeComponent<T> extends OwlDateTime<T>
                 .pipe(take(1))
                 .subscribe(() => {
                     this.popupRef.updatePosition();
+                });
+
+            this.pickerBeforeOpenedStreamSub = this.pickerContainer.beforePickerOpenedStream
+                .pipe(take(1))
+                .subscribe(() => {
+                    this.beforePickerOpen.emit(null);
                 });
 
             // emit open stream
