@@ -1,12 +1,12 @@
 /**
  * date-time.class
  */
-import { EventEmitter, Inject, Input, Optional } from '@angular/core';
+import {EventEmitter, Inject, Input, Optional, Directive} from '@angular/core';
 import {
     coerceBooleanProperty,
     coerceNumberProperty
 } from '@angular/cdk/coercion';
-import { DateTimeAdapter } from './adapter/date-time-adapter.class';
+import {DateTimeAdapter} from './adapter/date-time-adapter.class';
 import {
     OWL_DATE_TIME_FORMATS,
     OwlDateTimeFormats
@@ -20,6 +20,15 @@ export type PickerMode = 'popup' | 'dialog' | 'inline';
 
 export type SelectMode = 'single' | 'range' | 'rangeFrom' | 'rangeTo';
 
+export enum DateView {
+    MONTH = 'month',
+    YEAR = 'year',
+    MULTI_YEARS = 'multi-years'
+}
+
+export type DateViewType = DateView.MONTH | DateView.YEAR | DateView.MULTI_YEARS;
+
+@Directive()
 export abstract class OwlDateTime<T> {
     /**
      * Whether to show the second's timer
@@ -51,7 +60,20 @@ export abstract class OwlDateTime<T> {
      * The view that the calendar should start in.
      */
     @Input()
-    startView: 'month' | 'year' | 'multi-years' = 'month';
+    startView: DateViewType = DateView.MONTH;
+
+
+    /**
+     * Whether to should only the year and multi-year views.
+     */
+    @Input()
+    yearOnly = false;
+
+    /**
+     * Whether to should only the multi-year view.
+     */
+    @Input()
+    multiyearOnly = false;
 
     /**
      * Hours to change per step
@@ -95,16 +117,16 @@ export abstract class OwlDateTime<T> {
     /**
      * Set the first day of week
      */
-    private _firstDayOfWeek = 0;
+    private _firstDayOfWeek: number;
     @Input()
     get firstDayOfWeek() {
         return this._firstDayOfWeek;
     }
 
     set firstDayOfWeek(value: number) {
-        value = coerceNumberProperty(value, 0);
+        value = coerceNumberProperty(value);
         if (value > 6 || value < 0) {
-            this._firstDayOfWeek = 0;
+            this._firstDayOfWeek = undefined;
         } else {
             this._firstDayOfWeek = value;
         }
@@ -123,7 +145,7 @@ export abstract class OwlDateTime<T> {
         this._hideOtherMonths = coerceBooleanProperty(val);
     }
 
-    private _id: string;
+    private readonly _id: string;
     get id(): string {
         return this._id;
     }
@@ -142,6 +164,8 @@ export abstract class OwlDateTime<T> {
 
     abstract get startAt(): T | null;
 
+    abstract get endAt(): T | null;
+
     abstract get opened(): boolean;
 
     abstract get pickerMode(): PickerMode;
@@ -158,9 +182,13 @@ export abstract class OwlDateTime<T> {
 
     abstract monthSelected: EventEmitter<T>;
 
+    abstract dateSelected: EventEmitter<T>;
+
     abstract selectYear(normalizedYear: T): void;
 
     abstract selectMonth(normalizedMonth: T): void;
+
+    abstract selectDate(normalizedDate: T): void;
 
     get formatString(): string {
         return this.pickerType === 'both'
@@ -179,7 +207,7 @@ export abstract class OwlDateTime<T> {
             (!this.dateTimeFilter || this.dateTimeFilter(dateTime)) &&
             (!this.minDateTime ||
                 this.dateTimeAdapter.compare(dateTime, this.minDateTime) >=
-                    0) &&
+                0) &&
             (!this.maxDateTime ||
                 this.dateTimeAdapter.compare(dateTime, this.maxDateTime) <= 0)
         );
@@ -189,7 +217,7 @@ export abstract class OwlDateTime<T> {
         return false;
     }
 
-    constructor(
+    protected constructor(
         @Optional() protected dateTimeAdapter: DateTimeAdapter<T>,
         @Optional()
         @Inject(OWL_DATE_TIME_FORMATS)
@@ -198,16 +226,16 @@ export abstract class OwlDateTime<T> {
         if (!this.dateTimeAdapter) {
             throw Error(
                 `OwlDateTimePicker: No provider found for DateTimeAdapter. You must import one of the following ` +
-                    `modules at your application root: OwlNativeDateTimeModule, OwlMomentDateTimeModule, or provide a ` +
-                    `custom implementation.`
+                `modules at your application root: OwlNativeDateTimeModule, OwlMomentDateTimeModule, or provide a ` +
+                `custom implementation.`
             );
         }
 
         if (!this.dateTimeFormats) {
             throw Error(
                 `OwlDateTimePicker: No provider found for OWL_DATE_TIME_FORMATS. You must import one of the following ` +
-                    `modules at your application root: OwlNativeDateTimeModule, OwlMomentDateTimeModule, or provide a ` +
-                    `custom implementation.`
+                `modules at your application root: OwlNativeDateTimeModule, OwlMomentDateTimeModule, or provide a ` +
+                `custom implementation.`
             );
         }
 
@@ -216,7 +244,7 @@ export abstract class OwlDateTime<T> {
 
     protected getValidDate(obj: any): T | null {
         return this.dateTimeAdapter.isDateInstance(obj) &&
-            this.dateTimeAdapter.isValid(obj)
+        this.dateTimeAdapter.isValid(obj)
             ? obj
             : null;
     }
