@@ -21,6 +21,7 @@ import { NumberFixedLenPipe } from './numberedFixLen.pipe';
 import { OwlCalendarComponent } from './calendar.component';
 import { OwlDateTimeInlineComponent } from './date-time-inline.component';
 import { OwlDialogModule } from '../dialog/dialog.module';
+import { optionsProviders } from './options-provider';
 
 @NgModule({
     imports: [CommonModule, OverlayModule, OwlDialogModule, A11yModule],
@@ -53,9 +54,7 @@ import { OwlDialogModule } from '../dialog/dialog.module';
     providers: [
         OwlDateTimeIntl,
         OWL_DTPICKER_SCROLL_STRATEGY_PROVIDER,
-    ],
-    entryComponents: [
-        OwlDateTimeContainerComponent,
+        ...optionsProviders,
     ]
 })
 export class OwlDateTimeModule {
