@@ -21,101 +21,198 @@ export interface DialogPosition {
     right?: string;
 }
 
-export class OwlDialogConfig {
+export interface OwlDialogConfigInterface {
     /**
      * ID of the element that describes the dialog.
      */
-    public ariaDescribedBy?: string | null = null;
+    ariaDescribedBy?: string | null;
 
     /**
      * Whether to focus the dialog when the dialog is opened
      */
-    public autoFocus? = true;
+    autoFocus?: boolean;
 
     /** Whether the dialog has a backdrop. */
-    public hasBackdrop? = true;
+    hasBackdrop?: boolean;
 
     /**
      * Custom style for the backdrop
      * */
-    public backdropStyle?: any;
+    backdropStyle?: any;
 
     /** Data being injected into the child component. */
-    public data?: any = null;
+    data?: any ;
 
     /** Whether the user can use escape or clicking outside to close a modal. */
-    public disableClose? = false;
+    disableClose?: boolean;
 
     /**
      * ID for the modal. If omitted, a unique one will be generated.
      */
-    public id?: string;
+    id?: string;
 
     /**
      * The ARIA role of the dialog element.
      */
-    public role?: 'dialog' | 'alertdialog' = 'dialog';
+    role?: 'dialog' | 'alertdialog';
 
     /**
      * Custom class for the pane
      * */
-    public paneClass?: string | string[] = '';
+    paneClass?: string | string[];
 
     /**
      * Mouse Event
      * */
-    public event?: MouseEvent = null;
+    event?: MouseEvent;
 
     /**
      * Custom class for the backdrop
      * */
-    public backdropClass?: string | string[] = '';
+    backdropClass?: string | string[];
 
     /**
      * Whether the dialog should close when the user goes backwards/forwards in history.
      * */
-    public closeOnNavigation?: boolean = true;
+    closeOnNavigation?: boolean;
 
     /** Width of the dialog. */
-    public width?: string = '';
+    width?: string ;
 
     /** Height of the dialog. */
-    public height?: string = '';
+    height?: string;
 
     /**
      * The min-width of the overlay panel.
      * If a number is provided, pixel units are assumed.
      * */
-    public minWidth?: number | string;
+    minWidth?: number | string;
 
     /**
      * The min-height of the overlay panel.
      * If a number is provided, pixel units are assumed.
      * */
-    public minHeight?: number | string;
+    minHeight?: number | string;
 
     /**
      * The max-width of the overlay panel.
      * If a number is provided, pixel units are assumed.
      * */
-    public maxWidth?: number | string = '85vw';
+    maxWidth?: number | string;
 
     /**
      * The max-height of the overlay panel.
      * If a number is provided, pixel units are assumed.
      * */
-    public maxHeight?: number | string;
+    maxHeight?: number | string;
 
     /** Position overrides. */
-    public position?: DialogPosition;
+    position?: DialogPosition;
 
     /**
      * The scroll strategy when the dialog is open
      * Learn more this from https://material.angular.io/cdk/overlay/overview#scroll-strategies
      * */
-    public scrollStrategy?: ScrollStrategy = new NoopScrollStrategy();
+    scrollStrategy?: ScrollStrategy;
 
-    public viewContainerRef?: ViewContainerRef;
+    viewContainerRef?: ViewContainerRef;
+}
+
+export class OwlDialogConfig implements OwlDialogConfigInterface {
+    /**
+     * ID of the element that describes the dialog.
+     */
+    public ariaDescribedBy: string | null = null;
+
+    /**
+     * Whether to focus the dialog when the dialog is opened
+     */
+    public autoFocus = true;
+
+    /** Whether the dialog has a backdrop. */
+    public hasBackdrop = true;
+
+    /**
+     * Custom style for the backdrop
+     * */
+    public backdropStyle: any;
+
+    /** Data being injected into the child component. */
+    public data: any = null;
+
+    /** Whether the user can use escape or clicking outside to close a modal. */
+    public disableClose = false;
+
+    /**
+     * ID for the modal. If omitted, a unique one will be generated.
+     */
+    public id: string;
+
+    /**
+     * The ARIA role of the dialog element.
+     */
+    public role: 'dialog' | 'alertdialog' = 'dialog';
+
+    /**
+     * Custom class for the pane
+     * */
+    public paneClass: string | string[] = '';
+
+    /**
+     * Mouse Event
+     * */
+    public event: MouseEvent = null;
+
+    /**
+     * Custom class for the backdrop
+     * */
+    public backdropClass: string | string[] = '';
+
+    /**
+     * Whether the dialog should close when the user goes backwards/forwards in history.
+     * */
+    public closeOnNavigation = true;
+
+    /** Width of the dialog. */
+    public width = '';
+
+    /** Height of the dialog. */
+    public height = '';
+
+    /**
+     * The min-width of the overlay panel.
+     * If a number is provided, pixel units are assumed.
+     * */
+    public minWidth: number | string;
+
+    /**
+     * The min-height of the overlay panel.
+     * If a number is provided, pixel units are assumed.
+     * */
+    public minHeight: number | string;
+
+    /**
+     * The max-width of the overlay panel.
+     * If a number is provided, pixel units are assumed.
+     * */
+    public maxWidth: number | string = '85vw';
+
+    /**
+     * The max-height of the overlay panel.
+     * If a number is provided, pixel units are assumed.
+     * */
+    public maxHeight: number | string;
+
+    /** Position overrides. */
+    public position: DialogPosition;
+
+    /**
+     * The scroll strategy when the dialog is open
+     * Learn more this from https://material.angular.io/cdk/overlay/overview#scroll-strategies
+     * */
+    public scrollStrategy: ScrollStrategy = new NoopScrollStrategy();
+
+    public viewContainerRef: ViewContainerRef;
 
     constructor() {
         this.id = `owl-dialog-${uniqueId++}`;
