@@ -13,7 +13,7 @@ import {
     TemplateRef
 } from '@angular/core';
 import { Location } from '@angular/common';
-import { OwlDialogConfig } from './dialog-config.class';
+import { OwlDialogConfig, OwlDialogConfigInterface } from './dialog-config.class';
 import { OwlDialogRef } from './dialog-ref.class';
 import { OwlDialogContainerComponent } from './dialog-container.component';
 import { extendObject } from '../utils';
@@ -55,8 +55,8 @@ export const OWL_DIALOG_SCROLL_STRATEGY_PROVIDER = {
     useFactory: OWL_DIALOG_SCROLL_STRATEGY_PROVIDER_FACTORY
 };
 
-/** I
- * njection token that can be used to specify default dialog options.
+/**
+ * Injection token that can be used to specify default dialog options.
  * */
 export const OWL_DIALOG_DEFAULT_OPTIONS = new InjectionToken<OwlDialogConfig>(
     'owl-dialog-default-options'
@@ -67,6 +67,7 @@ export class OwlDialogService {
     private ariaHiddenElements = new Map<Element, string | null>();
 
     private _openDialogsAtThisLevel: OwlDialogRef<any>[] = [];
+    private _beforeOpenAtThisLevel = new Subject<OwlDialogRef<any>>();
     private _afterOpenAtThisLevel = new Subject<OwlDialogRef<any>>();
     private _afterAllClosedAtThisLevel = new Subject<void>();
 
@@ -75,6 +76,13 @@ export class OwlDialogService {
         return this.parentDialog
             ? this.parentDialog.openDialogs
             : this._openDialogsAtThisLevel;
+    }
+
+    /** Stream that emits when a dialog has been opened. */
+    get beforeOpen(): Subject<OwlDialogRef<any>> {
+        return this.parentDialog
+            ? this.parentDialog.beforeOpen
+            : this._beforeOpenAtThisLevel;
     }
 
     /** Stream that emits when a dialog has been opened. */
@@ -103,7 +111,7 @@ export class OwlDialogService {
                 : this._afterAllClosed.pipe(startWith(undefined))
     );
 
-    private scrollStrategy: () => ScrollStrategy;
+    private readonly scrollStrategy: () => ScrollStrategy;
 
     constructor(
         private overlay: Overlay,
@@ -112,7 +120,7 @@ export class OwlDialogService {
         @Inject(OWL_DIALOG_SCROLL_STRATEGY) scrollStrategy: any,
         @Optional()
         @Inject(OWL_DIALOG_DEFAULT_OPTIONS)
-        private defaultOptions: OwlDialogConfig,
+        private defaultOptions: OwlDialogConfigInterface,
         @Optional()
         @SkipSelf()
         private parentDialog: OwlDialogService,
@@ -126,7 +134,7 @@ export class OwlDialogService {
 
     public open<T>(
         componentOrTemplateRef: ComponentType<T> | TemplateRef<T>,
-        config?: OwlDialogConfig
+        config?: OwlDialogConfigInterface
     ): OwlDialogRef<any> {
         config = applyConfigDefaults(config, this.defaultOptions);
 
@@ -155,6 +163,7 @@ export class OwlDialogService {
         dialogRef
             .afterClosed()
             .subscribe(() => this.removeOpenDialog(dialogRef));
+        this.beforeOpen.next(dialogRef);
         this.afterOpen.next(dialogRef);
         return dialogRef;
     }
@@ -182,7 +191,7 @@ export class OwlDialogService {
         componentOrTemplateRef: ComponentType<T> | TemplateRef<T>,
         dialogContainer: OwlDialogContainerComponent,
         overlayRef: OverlayRef,
-        config: OwlDialogConfig
+        config: OwlDialogConfigInterface
     ) {
         const dialogRef = new OwlDialogRef<T>(
             overlayRef,
@@ -220,7 +229,7 @@ export class OwlDialogService {
     }
 
     private createInjector<T>(
-        config: OwlDialogConfig,
+        config: OwlDialogConfigInterface,
         dialogRef: OwlDialogRef<T>,
         dialogContainer: OwlDialogContainerComponent
     ) {
@@ -240,14 +249,14 @@ export class OwlDialogService {
         );
     }
 
-    private createOverlay(config: OwlDialogConfig): OverlayRef {
+    private createOverlay(config: OwlDialogConfigInterface): OverlayRef {
         const overlayConfig = this.getOverlayConfig(config);
         return this.overlay.create(overlayConfig);
     }
 
     private attachDialogContainer(
         overlayRef: OverlayRef,
-        config: OwlDialogConfig
+        config: OwlDialogConfigInterface
     ): OwlDialogContainerComponent {
         const containerPortal = new ComponentPortal(
             OwlDialogContainerComponent,
@@ -261,7 +270,7 @@ export class OwlDialogService {
         return containerRef.instance;
     }
 
-    private getOverlayConfig(dialogConfig: OwlDialogConfig): OverlayConfig {
+    private getOverlayConfig(dialogConfig: OwlDialogConfigInterface): OverlayConfig {
         const state = new OverlayConfig({
             positionStrategy: this.overlay.position().global(),
             scrollStrategy:
@@ -314,7 +323,7 @@ export class OwlDialogService {
             const siblings = overlayContainer.parentElement.children;
 
             for (let i = siblings.length - 1; i > -1; i--) {
-                let sibling = siblings[i];
+                const sibling = siblings[i];
 
                 if (
                     sibling !== overlayContainer &&
@@ -340,8 +349,8 @@ export class OwlDialogService {
  * @returns The new configuration object.
  */
 function applyConfigDefaults(
-    config?: OwlDialogConfig,
-    defaultOptions?: OwlDialogConfig
+    config?: OwlDialogConfigInterface,
+    defaultOptions?: OwlDialogConfigInterface
 ): OwlDialogConfig {
     return extendObject(new OwlDialogConfig(), config, defaultOptions);
 }
