@@ -3,13 +3,50 @@
  */
 
 export { OwlDateTimeModule } from './lib/date-time/date-time.module';
+
 export { OwlDateTimeIntl } from './lib/date-time/date-time-picker-intl.service';
+
 export { OwlNativeDateTimeModule } from './lib/date-time/adapter/native-date-time.module';
+
 export {
   OWL_DATE_TIME_LOCALE_PROVIDER,
   OWL_DATE_TIME_LOCALE,
-  DateTimeAdapter
+  DateTimeAdapter,
+
 } from './lib/date-time/adapter/date-time-adapter.class';
+
 export { OWL_DATE_TIME_FORMATS, OwlDateTimeFormats } from './lib/date-time/adapter/date-time-format.class';
+
+export {
+  UnixTimestampDateTimeAdapter
+} from './lib/date-time/adapter/unix-timestamp-adapter/unix-timestamp-date-time-adapter.class';
+
+export { OWL_UNIX_TIMESTAMP_DATE_TIME_FORMATS } from './lib/date-time/adapter/unix-timestamp-adapter/unix-timestamp-date-time-format.class';
+
 export { OwlDateTimeInlineComponent } from './lib/date-time/date-time-inline.component';
+
 export { OwlDateTimeComponent } from './lib/date-time/date-time-picker.component';
+
+export * from './lib/date-time/calendar-body.component';
+
+export * from './lib/date-time/timer.component';
+
+export * from './lib/date-time/date-time-picker-trigger.directive';
+
+export * from './lib/date-time/date-time-picker-input.directive';
+
+export * from './lib/date-time/calendar-multi-year-view.component';
+
+export * from './lib/date-time/calendar-year-view.component';
+
+export * from './lib/date-time/calendar-month-view.component';
+
+export * from './lib/date-time/calendar.component';
+
+export * from './lib/date-time/timer.component';
+
+export { NativeDateTimeAdapter } from './lib/date-time/adapter/native-date-time-adapter.class';
+
+export * from './lib/date-time/options-provider';
+
+export { PickerType, PickerMode, SelectMode, DateView, DateViewType } from './lib/date-time/date-time.class'
