@@ -42,8 +42,8 @@ export function createKeyboardEvent(
     target?: Element,
     key?: string
 ) {
-    let event = document.createEvent('KeyboardEvent') as any;
-    let originalPreventDefault = event.preventDefault;
+    const event = document.createEvent('KeyboardEvent') as any;
+    const originalPreventDefault = event.preventDefault;
 
     // Firefox does not support `initKeyboardEvent`, but supports `initKeyEvent`.
     if (event.initKeyEvent) {
