@@ -14,18 +14,20 @@ export class AppComponent implements AfterViewInit {
   @ViewChild('date_range_component', { static: true })
   date_range_component: OwlDateTimeComponent<AppComponent>;
   public selectedMoments: Moment[] = [
-    moment('2019-03-11T08:00:00+11:00').tz('Australia/Sydney'),
-    moment('2019-03-11T15:00:00+11:00').tz('Australia/Sydney')
+    moment('2019-03-11T08:00:00+11:00').tz('America/Los_Angeles'),
+    moment('2019-03-11T15:00:00+11:00').tz('America/Los_Angeles')
   ];
+
+
+  currentValue: Date = new Date('4/21/2020, 12:00 AM');
+  endValue: Date = new Date('4/21/2020, 11:59 PM');
 
   open_once = false;
 
   ngAfterViewInit() {
-    /*
-    if (!this.open_once) {
-        this.date_range_component.open();
-        this.open_once = true;
-    }
-    */
+  }
+
+  selectedTrigger(_date) {
+    console.log(_date);
   }
 }
